### 2020-07-18 #85 @whedon
[PRE REVIEW]: An Applied Geographic Information Systems and Science Course in R
**Submitting author:** @andrewmaclachlan (<a href="http://orcid.org/0000-0003-0356-1623">Andrew MacLachlan</a>)
**Repository:** <a href="https://github.com/andrewmaclachlan/CASA0005repo" target ="_blank">https://github.com/andrewmaclachlan/CASA0005repo</a>
**Version:** 2.0
**Reviewers:** tbeuzen

@andrewmaclachlan if you have any suggestions for potential reviewers then please mention them here in this thread (without tagging them with an @). In addition, [this list of people](https://bit.ly/jose-reviewers) have already agreed to review for JOSE and may be suitable for this submission.
