### 2020-07-18 #85 @labarba
@openjournals/dev Can you help us find out what is causing these errors? The [paper.md](https://github.com/andrewmaclachlan/CASA0005repo/blob/master/paper.md) file _is there_.
