### 2020-07-27 #85 @arfon
I think Whedon is timing out here because of the very large size of this repository and therefore reporting it can't find the paper (apologies for the unhelpful error message).

@andrewmaclachlan - your repository is ~1.4Gb (> 3Gb on disk) which means Whedon is going to struggle with this repository (as will your users). Is there some way you could reduce the size of the repository?
