### 2020-07-28 #85 @andrewmaclachlan
Hi there, looks like Weedon did run (i assume it just creates the pdf proof) and that seems fine to me. I tried to stored some satellite imagery using GitHub LFS, however, it didn't work but means i have to delete the repo to remove them fully. I can do this if needed.
