+++ paper.md
@@ -1,0 +1,69 @@
+---
+title: 'An Applied Geographic Information Systems and Science Course in R'
+tags:
+- Spatial data science
+- Smart cities
+- rstats
+authors:
+- name: Andrew MacLachlan
+  orcid: 0000-0003-0356-1623
+  affiliation: "1"
+- name: Adam Dennett
+  orcid: 0000-0001-8036-0185
+  affiliation: "1"
+affiliations:
+- name: The Bartlett Centre for Advanced Spatial Analysis, University College London
+  index: 1
+date: 05 May 2020
+bibliography: paper.bib
+---
+
+# Summary
+
+This paper presents a masters level geographic information systems and science course (CASA0005) developed by the Bartlett Centre for Advanced Spatial Analysis (CASA), University College London (UCL). CASA0005 is a compulsory module for both MSc programs within the department, running annually in the autumn term with between 100 and 150 enrolled students. During the summer of 2019 the module was transformed from static practicals with some legacy content related to commercial software to an online interactive book primarily for the open source R data science programming language written using the Bookdown package [@Xie2016].
+
+# Statement of need
+
+The buzzword ‘Big data’ and associated roles of data scientists have become common additions to academia, industry and government organisations seeking to optimise workflows and base decisions on factual evidence [@Donoho2017; @Li2016]. In response, the demand for data scientists has significantly grown, with two major recruitment websites LinkedIn and Glassdoor reporting a respective 56% and 44% year on year increase in advertised roles along with classification as the most promising career based on salary, growth, promotion opportunities and satisfaction within the USA in 2019 [@Glassdoor2020; @Pattabiraman2019]. In a similar theme data from the UK’s Universities and Colleges Admissions Service (UCAS) shows that the number of postgraduate level data science courses within the UK has grown from 104 across 51 institutions to 195 across 85 institutions between the academic years commencing in 2019 and 2020 [@TheUniversitiesandCollegesAdmissionsService2020]
+. However, for the ability to use data to it’s full capability the spatial aspect is an essential component, yet the majority of current academic resources provide limited coverage of advanced, applied and reproducible geospatial methodologies using unfiltered and real world data examples [@Arribas-Bel2019]
+
+Whilst materials such as the R for Data Science book provide a foundational and essential resource for learning the R data science programming language, consideration of spatial data is extremely limited [@Grolemund2017]. The welcome addition of Geocomputation with R specifically addressed this resource gap, but often used basic examples that are unrepresentative of the complex spatial data wrangling, transformation and manipulation skills required by masters level graduates [@Lovelace2019]. Consequently, this paper presents an academic teaching resource using applied real world government (e.g. [@Cosh2020]) and academic examples (e.g. [@Guha2018]) manipulating complex and large datasets in creating reproducible and actionable outputs, being made possible through delivery as a taught masters level module. 
+
+# Learning objectives
+
+Learning objectives are provided at the start of each individual practical that combine to achieve the overall module learning objectives.
+
+After having taking this module, students should be able to:
+
+* Develop a working knowledge of QGIS and R to support the application of GI Science techniques
+* Visualise geographic information through producing appropriate maps to high cartographic standards
+* Carry out spatial data management tasks (joining attribute to geometry data, cleaning data, converting between file formats and spatial reference systems)
+* Interpret data and apply relevant spatial analyses (e.g. auto correlation/hot spot analysis, areal interpolation, point in polygon/buffer analysis, spatial statistical analysis) to answer a variety of spatial problems
+* Explain and evaluate common issues with geographic data such as representation and uncertainty
+* Apply and critique (spatial) statistical analysis techniques to infer relationships between spatial phenomena
+* Experience the diversity of the global spatial data landscape and evaluate the relative drawbacks and merits of different spatial datasets
+
+# Content
+
[…]
