+++ paper.bib
@@ -1,0 +1,242 @@
+@article{Arribas-Bel2019,
+author = {Arribas-Bel, Dani},
+doi = {10.21105/jose.00042},
+file = {:C$\backslash$:/Users/Andy/Downloads/10.21105.jose.00042.pdf:pdf},
+journal = {Journal of Open Source Education},
+number = {14},
+pages = {42},
+title = {{A course on Geographic Data Science}},
+volume = {2},
+year = {2019}
+}
+@article{Artis1982,
+abstract = {This study investigates the effects of roof emissivity variation of aerial thermogram images. Thermograms have been used to detect heat loss from residential roofs. Emissivity variation among rooftops, however, can lead to a misrepresentation of the temperatures mapped in a thermogram image. The objectives of this study were (a) to demonstrate the feasibility of a technique to use remotely sensed data to calculate surface emissivities and (b) to apply that technique, to determine the extent of emissivity variation in urban surfaces. In the first part of the experimental approach a passive technique is developed to calculate emissivity from two-band thermal infrared radiance data. Inherent limitations and sources of error associated with the technique are discussed. In the second part of the experimental approach the technique was used to measure the emissivity of 1411 roofs within the city limits of Terre Haute, Indiana. Results of this survey indicated that over 98{\%} of the roofs surveyed were confined to a very narrow range of emissivities. It is concluded that the observed variation in rooftop emissivities has a minimal effect on the temperatures depicted in thermograms. ?? 1982.},
+author = {Artis, David A. and Carnahan, Walter H.},
+doi = {10.1016/0034-4257(82)90043-8},
+file = {:C$\backslash$:/Users/Andy/AppData/Local/Mendeley Ltd./Mendeley Desktop/Downloaded/Artis, Carnahan - 1982 - Survey of emissivity variability in thermography of urban areas.pdf:pdf},
+isbn = {0034-4257},
+issn = {00344257},
+journal = {Remote Sensing of Environment},
+number = {4},
+pages = {313--329},
+title = {{Survey of emissivity variability in thermography of urban areas}},
+volume = {12},
+year = {1982}
+}
+@article{Brunsdon1996,
+abstract = {Geographically weighted regression (GWR) is a local form of spatial analysis introduced in 1996 in the geographical literature drawing from statistical approaches for curve-fitting and smoothing applications. The method works based on the simple yet powerful idea of estimating local models using subsets of observations centered on a focal point. Since its introduction, GWR rapidly captured the attention of many in geography and other fields for its potential to investigate nonstationary relations in regression analysis. The basic concepts have also been used to obtain local descriptive statistics and other models such as Poisson regression and probit. The method has been instrumental in highlighting the existence of potentially complex spatial relationships. At the same time, there have been a number of debates concerning the nature and range of applications of the method, including its use for inferential analysis or interpolation. The evidence available suggests that GWR is a useful, if imperfect, tool for inferring spatial processes, and a relatively simple and effective tool for spatial interpolation. Related technical developments enhance GWR (e.g., autocorrelation tests and multiple comparison adjustments) and/or complement it (e.g., the expansion method). Other approaches provide alternatives to the use of GWR (e.g., kriging and Bayesian models).},
+author = {Brunsdon, Chris and Fotheringham, A. Stewart and Charlton, Martin E.},
+doi = {10.1016/B978-008044910-4.00447-8},
+file = {:C$\backslash$:/Users/Andy/Downloads/j.1538-4632.1996.tb00936.x.pdf:pdf},
+isbn = {9780080449104},
+journal = {Geographical analysis},
[…]
