+++ _test.Rmd
@@ -1,0 +1,19 @@
+# test
+
+```{r interactive, fig.cap="The blue markers indicate where the authors are from. The basemap is a tiled image of the Earth at night provided by NASA. Interact with the online version at geocompr.robinlovelace.net, for example by zooming in and clicking on the popups.", out.width="100%", fig.scap="Where the authors are from.", echo=FALSE}
+if(knitr::is_latex_output()){
+    knitr::include_graphics("interactive.png")   
+} else if(knitr::is_html_output()){
+     #library(leaflet)
+     #popup = c("Robin", "Jakub", "Jannes")
+     #interactive = leaflet() %>%
+      # addProviderTiles("NASAGIBS.ViirsEarthAtNight2012") %>%
+       #addMarkers(lng = c(-3, 23, 11),
+        #          lat = c(52, 53, 49), 
+         #         popup = popup)  
+     #library(htmlwidgets)
+     #saveWidget(interactive, file = "interactive.html")
+     #file.copy("interactive.html", "Prac1_data/interactive.html")
+    knitr::include_url("Prac1_data/interactive.html")
+}
+```

[…]
