+++ _10-prac10.Rmd
@@ -1,1114 +1,0 @@
-# Advanced R, MAUP and more regression
-
-## Learning outcomes
-
-1. Describe, explain and visualise the MAUP problem
-1. Design and use loops and functions
-1. Execute linear, Ridge and LASSO regression to predict values (e.g. future or missing)
-1. Critically evaluate different regression approaches 
-
-## Homework 
-
-Outside of our schedulded sessions you should be doing around 12 hours of extra study per week. Feel free to follow your own GIS interests, but good places to start include the following:
-
-::: {.infobox .assignment data-latex="{note}"}
-**Assignment**
-
-From weeks 10 until the assignment deadline, undertake your analysis, then write up your discussion and conclusion **(that includes recommendations)**.
-:::
-
-::: {.infobox .tip data-latex="{note}"}
-**Before submission**
-
-Before submission, check that your assignment follows the standard model of scientific investigation. You may need to update part 1 based on the analysis you undertook to ensure a seamless narrative throughout. 
-:::
-
-::: {.infobox .note data-latex="{note}"}
-**Reading**
-This week:
-
-* For advanced R coding [Chapter 5 "Control flow"](https://adv-r.hadley.nz/control-flow.html) from Advacned R by Hadley Wickham.
-
-* For more advanced R inclduing functions [Chapter 10 "Scripts, algorithms and functions"](https://geocompr.robinlovelace.net/algorithms.html) from Geocomputation with R by Lovelace, Nowosad and Muenchow (2020).
-
-* For Ridge and LASSO regression [Chapter 4 "Linear models III: shrinkage, multivariate response, and big data"](https://bookdown.org/egarpor/PM-UC3M/lm-iii.html) from Predictive Modeling by Portugués (2020).
-
-Remember this is just a starting point, explore the [reading list](https://rl.talis.com/3/ucl/lists/139FBAF8-DACD-60FB-8BDC-E9C9E09BA885.html?lang=en-GB&login=1), practical and lecture for more ideas.
-:::
-
-## Recommended listening `r emo::ji("headphones")`
-
-Some of these practicals are long, take regular breaks and have a listen to some of our fav tunes each week.
-
-[Andy](https://www.youtube.com/watch?v=YNXybY2DtwM)
-
-[Adam](https://open.spotify.com/album/5yMqI30BaFZww5orpqemV5?si=vgEFH5p9ReqtGHSt66vCsw)6 letters: LSB, DRS. Oh yes!
-
-## Introduction
-
-The Modifiable Areal Unit Problem (MAUP) represents the related effects of scale and aggregation on all geographic data and analysis. It was first deomstrated by the geographer Stan Openshaw in 1984 who showed that as you aggregated results to diffrent spatial units the reults could be manipulated to show different outcomes. Throughout this practical book we've considered London boroughs, but would any of the results change if we considered the data at ward level? Are we hiding any trends because we have just used the borough data that summed all of the wards within the borough? It's important to consider what is the most appropraite spatial unit for your analysis and provide appropraite reasoning. Even this is pretty straightforward (e.g. the data was only provided at the borough level) you must contextualise this and describe any potential limitations of it. In this practical i will firstly demonstrate the MAUP in action using some more advanced R code. Then we will have a look at some technqieus to model and validate data.
-
[…]
