### 2020-09-03 #85 @andrewmaclachlan
Hi @labarba. I've managed to significantly reduce the size of the repo and correct some of the missing references in the article proof. I'd be happy for any of the following to review if they are able to:

Karthik Ram - based on their reproducibility work in R
Matthew Forshaw - based on similar disciplines and uses R
Joseph Stachelek - based on geographical area of work and uses R

Please let me know if you need anymore suggestions. The repo and website are being updated over this term to fix any issues with delivery, however all of the content will remain the same.
