### 2020-09-03 f913920
all working including that interactive plot
