+++ 11-assignment.Rmd
@@ -6,30 +6,42 @@
 library(knitr)
 ```
 
+The following is taken from the assignment requirements sheet and mark scheme found on Moodle.
+
 ## Requirements
 
-**Deadline:** Monday 13th January 2020, 5pm. There are no opportunities to submit a second version after this deadline, this means it's a one time only submission.
+**Deadline:** Monday 11th January 2021, 5pm. There are no opportunities to submit a second version after this deadline.
 
-**Submission:** PDF via Moodle. If anything other than a .pdf is submitted **it will not be marked**. 
+Late submission of the assessment will be subject to the [standard penalties according to the UCL regulations](https://www.ucl.ac.uk/academic-manual/chapters/chapter-4-assessment-framework-taught-programmes/section-3-module-assessment#3.12)
 
-**Coversheet:** You must have a coversheet at the start of your assignment (within the single PDF) that includes your assignment title, programme, department, GitHub repository URL and any online website / instructions you have created (e.g. RPubs). There is no standard format for the cover sheet, just ensure it has all the required information. It is not included in the word count. **Check all of your URLs work (e.g. in a different browser)**.
+**Weighting:** This single assessment makes up 100% of the marks for the module.
 
-**Weighting:** This assessment makes up 100% of the marks for this module but is divided into two parts (worth 50% each) that must be submitted together in a single PDF document.
+**Submission:** PDF via Moodle. **Do not upload anything other than a PDF.**
+
+**Coversheet:** You must have a coversheet at the start of your assignment (within the single PDF) that includes your:
+* Assignment title
+* Programme
+* Department
+* GitHub repository URL and any online website / instructions you have created (e.g. RPubs). 
+
+There is no standard format for the cover sheet, just ensure it has all the required information. It is not included in the word count. Check all of your URLs work (e.g. in a different browser).
 
 **Assessment outline:**
 
-For the assessment of this course you will combine the cartographic and GIScience skills you have learned over the last 10 weeks to showcase your abilities as a fully rounded spatial analyst. Your task is to carry out a mini research project which answers a pertinent or topical geo-spatial question/hypothesis/issue in a logical, scientific and reproducible manner. Your project can be located anywhere on Earth and at any spatial scale (e.g. local, national, international) but must involve analysis of some appropriate spatial data (raster, vector or both) that you have sourced yourself. You must not reuse the data we provided in the practical sessions. You will submit only one document, but the assessment process will be divided into 2 parts detailed below, each worth 50% of the final grade and around 1500 words. The final total (of both parts combined) should not exceed 3000 words excluding references. The reference list should be provided on a new page within your PDF document. You must conduct your analysis using R and/or QGIS.
+For the assessment of this course you will combine the cartographic and GIScience skills you have learned over the last 10 weeks to showcase your abilities as a fully rounded spatial analyst. Your task is to carry out a mini research project which answers a pertinent or topical geo-spatial question/hypothesis/issue in a logical, scientific and reproducible manner. Your project can be located anywhere on Earth and at any spatial scale (e.g. local, national, international) but must involve analysis of some appropriate spatial data (raster, vector or both). You must not directly reuse the data we provided in the practical sessions (see FAQs below for further guidance). You will submit only one document that should not exceed 3000 words excluding references. **The reference list should be provided on a new page** within your PDF document.
 
-**Part 1**
+**Assessment task**
 
-In part 1 of the assessment you are required to construct a relevant spatial problem founded upon relevant and pertinent literature (e.g. academic and policy) and propose a methodology to test your hypothesis. In this section we expect to see an: introduction, research question (or hypothesis), literature review and methodology.
[…]
