+++ 01-prac1.Rmd
@@ -18,7 +18,7 @@ Outside of our schedulded sessions you should be doing around 12 hours of extra 
 ::: {.infobox .assignment data-latex="{note}"}
 **Assignment**
 
-This week you should undertake wider data exploration and reading to find a topic you might be interested in progressing with. I'd recommend starting with a data search first to avoid proposing a research area then forcing potentially unsuitable data to fit it.
+This week you should undertake wider data exploration and reading to find a topic you might be interested in progressing with. Start with a data search first to avoid proposing a research area then forcing potentially unsuitable data to fit it.
 :::
 
 ::: {.infobox .note data-latex="{note}"}

[…]
