+++ 12-assignment_resources.Rmd
@@ -30,13 +30,15 @@ The [UK ESRI conference](https://www.esriuk.com/en-gb/about/events/ac/overview) 
 
 New packages and functions that i've recently come across that are worth exploring..
 
-* RStudio cloud: https://rstudio.cloud/ --- like RStudio but online, great for collaborating
-* Animating plots: https://rpubs.com/jgleeson/ripple-effect
-* Bivariate maps: https://timogrossenbacher.ch/2019/04/bivariate-maps-with-ggplot2-and-sf/
-* Rayshader: https://www.rayshader.com/
-* Cartography package: http://riatelab.github.io/cartography/docs/
-* Data is beautiful reddit: https://www.reddit.com/r/dataisbeautiful/
-* Network visualisation: https://kateto.net/network-visualization
+<div class="col2">
+* [RStudio cloud](https://rstudio.cloud/) --- like RStudio but online, great for collaborating
+* [Animating plots](https://rpubs.com/jgleeson/ripple-effect)
+* [Bivariate maps](https://timogrossenbacher.ch/2019/04/bivariate-maps-with-ggplot2-and-sf/)
+* [Rayshader](https://www.rayshader.com/)
+* [R Cartography package](http://riatelab.github.io/cartography/docs/)
+* [Data is beautiful reddit](https://www.reddit.com/r/dataisbeautiful/)
+* [Network visualisation](https://kateto.net/network-visualization)
+</div><br/>
 
 ## Books/reading resources
 
@@ -55,7 +57,7 @@ There are a lot of free online books for geospatial analysis, especially using R
 * [Geocomputation with R ](https://geocompr.robinlovelace.net/)
 * [Happy Git and GitHub for the useR](https://happygitwithr.com/)
 * [Bookdown information pages](https://bookdown.org/)
-* [Interactive web-based data visualization with R, plotly, and shiny https://plotly-r.com/index.html](https://plotly-r.com/index.html)
+* [Interactive web-based data visualization with R, plotly, and shiny](https://plotly-r.com/index.html)
 * [YaRrr! The Pirate’s Guide to R](https://bookdown.org/ndphillips/YaRrr/)
 * [Engineering Production-Grade Shiny Apps](https://thinkr-open.github.io/building-shiny-apps-workflow/)
 * [Geospatial Health Data: Modeling and Visualization with R-INLA and Shiny](http://www.paulamoraga.com/book-geospatial/index.html)
@@ -71,20 +73,23 @@ If you want to produce more efficent R code then have a look at the [Efficient R
 ## New developments
 
 Every year there is a useR conference that provides a load to tutorials on the latest developments and research...check them out here: 
+<div class="col2">
+* [Data to viz, really great website with code](https://www.data-to-viz.com/)
+* [Modern data plots](https://moderndata.plot.ly/visualizing-geo-spatial-data-with-sf-and-plotly/)
+* [Make graphics like the BBC](https://github.com/bbc/bbplot/blob/master/chart_examples/bbplot_example_plots.png)
+* [UseR2019 tutorials](http://www.user2019.fr/tutorials/)
+* [UseR2019 materials](https://github.com/sowla/useR2019-materials)
+</div><br/>
 
-* LOOK AT THIS WEBSITE for data viz and code: https://www.data-to-viz.com/
-* ALSO this website for data viz: https://moderndata.plot.ly/visualizing-geo-spatial-data-with-sf-and-plotly/
-* Make graphics like the BBC: https://github.com/bbc/bbplot/blob/master/chart_examples/bbplot_example_plots.png
[…]
