### 2020-09-25 c9f791d
remove assingment info
