+++ 01-prac1.Rmd
@@ -41,9 +41,9 @@ This week:
 
 Some of these practicals are long, take regular breaks and have a listen to some of our fav tunes each week.
 
-[Andy](https://www.youtube.com/watch?v=1egwdCGi_dA&list=OLAK5uy_lGype0gR4vyIE0J2hvNYtpifSbNynJTNk&index=1)
+[Andy](https://open.spotify.com/album/1A3nVEWRJ8yvlPzawHI1pQ). Week 1. No other choice than Vampire Weekend. The band formed during college (University in the USA) and produced their first album of their own whilst working full time jobs! Incredible. The lead vocalist and guitarist Ezra Koenig was a school teacher in Brooklyn, New York. I've seen them a few times at Glastonbuty and in London during which they gave the people what they wanted and took requests from their entire catalogue, really amazing musicians. Note some of the guest appearances in this album as they also might make an apperance later in the term! 
 
-[Adam](https://open.spotify.com/album/0XXI8pxXeS9cvpBX9bB3hh)... This week, wrap your ears around this piece of liquid gold from Japan’s finest, Makoto.
+[Adam](https://open.spotify.com/album/3LVWk5ZlGgFPvTUVSo8EKh?si=0tvDERBmRAOQgyRztZK-iw) OK, it's week 1, so I'm going in HARD. This week it's the raver's raver, the DJ's DJ. This man is a style icon and so drum & bass it hurts - it's Voltage! Here he is with his new album, Balance Over Symmetry. This will surprise a lot of people as it's an eclectic mix of DnB bangers through to some Burial-inspired minimal dubstep. Follow him on [Instagram](https://www.instagram.com/voltage_uk_/) and get your ears around his brilliant.
 
 ## The Basics of Geographic Information
 
