### 2020-10-07 40b9b93
minor edits to Rstudio UCL server info
