### 2020-10-14 fca5f2a
minor updates to prac2
