### 2020-10-15 573291c
Update 02-prac2.Rmd
