+++ 02-prac2.Rmd
@@ -1106,6 +1106,11 @@ flytipping1 <- read_csv("https://data.london.gov.uk/download/fly-tipping-inciden
 # view the data
 view(flytipping1)
 ```
+Fixing the csv file
+```
+flytipping1 <- flytipping1 %>%
+  mutate(year,year=replace(year, year > "2018-19","2018-19"))
+```
 
 So we have a tibble with columns of each varaible (e.g. warning letters, total actions taken) where every row is a London borough. We want make sure that each observation has its own row...it doesn't as in the first row here we have observations for total incidents and total actions taken etc...to do this we will use `pivot_longer()`. Make sure you have the most recent version of `tidyverse` if you get an error.
 
