+++ 05-prac5.Rmd
@@ -319,13 +319,18 @@ newbb <- c(xmin=-296000, ymin=5408, xmax=655696, ymax=1000000)
 UK_outlinecrop <- UK_outline$geometry %>%
   st_crop(., newbb)
 
-#change the column name from hotels_in_borough for the legend
-
 Hotels <- Hotels %>%
-  dplyr::rename(`Accomodation count` = hotels_in_borough)
+  #at the moment each hotel is a row for the borough
+  #we just one one row that has number of airbnbs
+  group_by(., GSS_CODE, NAME)%>%
+  summarise(`Accomodation count` = unique(hotels_in_borough))
 
 Airbnb <- Airbnb %>%
-  dplyr::rename(`Accomodation count` = hotels_in_borough)
+  group_by(., GSS_CODE, NAME)%>%
+  summarise(`Accomodation count` = unique(hotels_in_borough))
+```
+Make the map
+```{r, warning=FALSE, message=FALSE, out.width = "800pt", cache=FALSE}
 
 tmap_mode("plot")
 
@@ -335,13 +340,17 @@ breaks = c(0, 5, 12, 26, 57, 286)
 
 # plot each map
 tm1 <- tm_shape(Hotels) + 
-  tm_polygons("Accomodation count", breaks=breaks)+
+  tm_polygons("Accomodation count", 
+              breaks=breaks,
+              palette="PuBu")+
   tm_legend(show=FALSE)+
   tm_layout(frame=FALSE)+
   tm_credits("(a)", position=c(0,0.85), size=1.5)
 
 tm2 <- tm_shape(Airbnb) + 
-  tm_polygons("Accomodation count", breaks=breaks) + 
+  tm_polygons("Accomodation count",
+              breaks=breaks, 
+              palette="PuBu") + 
   tm_legend(show=FALSE)+
   tm_layout(frame=FALSE)+
   tm_credits("(b)", position=c(0,0.85), size=1.5)
@@ -354,7 +363,8 @@ tm3 <- tm_shape(UK_outlinecrop)+
   tm_text("CITY_NAME", xmod=-1, ymod=-0.5)
 
 legend <- tm_shape(Hotels) +
-    tm_polygons("Accomodation count") +
+    tm_polygons("Accomodation count",
+                palette="PuBu") +
     tm_scale_bar(position=c(0.2,0.04), text.size=0.6)+
     tm_compass(north=0, position=c(0.65,0.6))+
     tm_layout(legend.only = TRUE, legend.position=c(0.2,0.25),asp=0.1)+
@@ -380,12 +390,61 @@ print(legend, vp=viewport(layout.pos.col=2, layout.pos.row=2, height=5))
 
 ```
 
+### Inset map
+
+```{r warning=FALSE, message=FALSE, out.width = "800pt", cache=TRUE}
+
+Londonbb = st_bbox(Airbnb,
+                    crs = st_crs(Airbnb)) %>% 
+  st_as_sfc()
+```
+
+
+```{r warning=FALSE, message=FALSE, out.width = "800pt", cache=TRUE}
+
+main <- tm_shape(Airbnb, bbbox = Londonbb) + 
+  tm_polygons("Accomodation count",
+              breaks=breaks, 
+              palette="PuBu")+
+  tm_scale_bar(position = c("left", "bottom"), text.size = .75)+
+  tm_layout(legend.position = c("right","top"), 
+            legend.text.size=.75, 
+            legend.title.size = 1.1,
+            frame=FALSE)+
+  tm_credits("(c) OpenStreetMap contrbutors and Air b n b", position=c(0.0,0.0))+
[…]
