+++ 03-prac3.Rmd
@@ -16,7 +16,12 @@ Outside of our schedulded sessions you should be doing around 12 hours of extra 
 ::: {.infobox .assignment data-latex="{note}"}
 **Assignment**
 
-This week you need to assess the validity of your topic in terms of data suitability, then sketch out an introduction and literature review including academic and policy documents (or any reputable source) --- it's the same as last week!
+This week you need to 
+
+* Assess the validity of your topic in terms of data suitability, then sketch out an introduction and literature review including academic and policy documents (or any reputable source) --- it's the same as last week!
+
+* Go through the assignment examples on Moodle and mark them with the mark scheme.
+
 :::
 
 ::: {.infobox .note data-latex="{note}"}
@@ -41,9 +46,9 @@ Remember this is just a starting point, explore the [reading list](https://rl.ta
 
 Some of these practicals are long, take regular breaks and have a listen to some of our fav tunes each week.
 
-[Andy](https://www.youtube.com/watch?v=nH7bjV0Q_44)
+[Andy]https://open.spotify.com/album/2P63KJNeC3DwaLpw8KHYqE). Oh Wonder, from London, an alt-pop duo according to Wikipedia. I just really like their music. Was due to see them in December in London, but it's been postponed. Ultralife is their first album and the one i prefer, tracks 2-4 are where it's at! 
 
-[Adam](https://open.spotify.com/album/23jTvoFSWLKhfS8BWIm12x?si=IsB-d7njTC-FAzLTZYzSwA) --- this week, from me, it’s a history lesson. 30 Years ago, two DJs started playing tunes in the other room in Heaven, just under Charing Cross Station. The night was called Rage and the DJs were called Fabio and Grooverider. Their mix of house and sped-up breakbeats was an entirely new sound that began to be called ‘Jungle’. The 30 Years of Rage album recalls some of the early tunes that helped shape an entire genre of music. Enjoy! 
+[Adam](https://open.spotify.com/album/4hEKRyUeVXejefnsvKYERB?si=FSgLYASmTh6lLNIcKXqtjg) OK, this week it's the 21st Century's answer to Shakespear - the man who provided the soundtrack to countless post-club 'back-to-your-house?' early mornings in my youth. Yes, straight outta Brum (via Brixton), he's only gone and penned a bunch of awesome new tracks almost 20 years after he first burst out of his bedroom studio. It's only flippin' Mike Skinner. You're listening to the Streets!
 
 ## Introduction 
 
