+++ 03-prac3.Rmd
@@ -46,7 +46,7 @@ Remember this is just a starting point, explore the [reading list](https://rl.ta
 
 Some of these practicals are long, take regular breaks and have a listen to some of our fav tunes each week.
 
-[Andy]https://open.spotify.com/album/2P63KJNeC3DwaLpw8KHYqE). Oh Wonder, from London, an alt-pop duo according to Wikipedia. I just really like their music. Was due to see them in December in London, but it's been postponed. Ultralife is their first album and the one i prefer, tracks 2-4 are where it's at! 
+[Andy](https://open.spotify.com/album/2P63KJNeC3DwaLpw8KHYqE). Oh Wonder, from London, an alt-pop duo according to Wikipedia. I just really like their music. Was due to see them in December in London, but it's been postponed. Ultralife is their first album and the one i prefer, tracks 2-4 are where it's at! 
 
 [Adam](https://open.spotify.com/album/4hEKRyUeVXejefnsvKYERB?si=FSgLYASmTh6lLNIcKXqtjg) OK, this week it's the 21st Century's answer to Shakespear - the man who provided the soundtrack to countless post-club 'back-to-your-house?' early mornings in my youth. Yes, straight outta Brum (via Brixton), he's only gone and penned a bunch of awesome new tracks almost 20 years after he first burst out of his bedroom studio. It's only flippin' Mike Skinner. You're listening to the Streets!
 
