+++ 03-prac3.Rmd
@@ -85,7 +85,7 @@ You might hear some key words about projections that could terrify you! Let's br
 * Datum = contains the point relationship (where the origin (0,0) of the map is) between a Cartesian coordinates (flat surface) and Earth's surface. They can be local or geocentric. They set the origin, the scale and orientation of the Coordiante Reference System (CRS).
 * Local datum = changes the Ellispoid to align with a certain location on the surface (e.g. BNG that uses the OSGB36 datum). A local datum is anything that isn't the centre of the Earth.
 * Geocentric datum =  the centre is equal to the Earth's centre of gravity (e.g. WGS84).
-* Coordinate reference system (CRS) = Formula that defines how the 2D map (e.g. on your screen or a paper map) relates to the 3D Earth. Also sometimes called a spatial Reference System (SRS)
+* Coordinate reference system (CRS) = Formula that defines how the 2D map (e.g. on your screen or a paper map) relates to the 3D Earth. Also sometimes called a spatial Reference System (SRS). It also stores the datum information.
 
 ::: {.infobox .tip data-latex="{note}"}
 
