### 2020-10-23 f2c674d
minor prac3 updates
