+++ packages.bib
@@ -27,8 +27,8 @@
   title = {rmarkdown: Dynamic Documents for R},
   author = {JJ Allaire and Yihui Xie and Jonathan McPherson and Javier Luraschi and Kevin Ushey and Aron Atkins and Hadley Wickham and Joe Cheng and Winston Chang and Richard Iannone},
   year = {2020},
-  note = {R package version 2.3},
-  url = {https://CRAN.R-project.org/package=rmarkdown},
+  note = {R package version 2.4.6},
+  url = {https://github.com/rstudio/rmarkdown},
 }
 
 @Book{bookdown2016,
@@ -71,5 +71,15 @@
   year = {2018},
   note = {ISBN 9781138359338},
   url = {https://bookdown.org/yihui/rmarkdown},
+}
+
+@Book{rmarkdown2020,
+  title = {R Markdown Cookbook},
+  author = {Yihui Xie and Christophe Dervieux and Emily Riederer},
+  publisher = {Chapman and Hall/CRC},
+  address = {Boca Raton, Florida},
+  year = {2020},
+  note = {ISBN 9780367563837},
+  url = {https://bookdown.org/yihui/rmarkdown-cookbook},
 }
 
