+++ 04-prac4.Rmd
@@ -195,7 +195,7 @@ Alternatively if you would like to use Git but you’re working on the UCL Remot
 
 1. In the Start Menu, open the git GUI. Start > Git > Git GUI. You should open the existing repository that you have just created.
 
-1. Whenever you have made some changes to your files in your cloned repo, you can use git to review the changes and ‘Commit’ (save) them and then ‘Push’ them up to your master repository on GitHub.
+1. Whenever you have made some changes to your files in your cloned repo, you can use git to review the changes and ‘Commit’ (save) them and then ‘Push’ them up to your main repository on GitHub.
 
 1. To review and commit your changes, in the commit menu, simply:
 
@@ -208,7 +208,7 @@ d. push the changes to your GitHub repo
 
 Now we can push our changes to GitHub using the up arrow either in the RStudio Git tab (envrionment quadrant), or from the review changes box (opens when you click commit).
 
-1. To do this, first make sure you have committed any changes to your local cloned repo and then click the ‘Push’ button to whizz your code up to your master GitHub repo --- you might be prompted to enter your github username and password to enable this...
+1. To do this, first make sure you have committed any changes to your local cloned repo and then click the ‘Push’ button to whizz your code up to your main GitHub repo --- you might be prompted to enter your github username and password to enable this...
 
 But....if the push button is greyed out go to the section [Greyed out push button]
 
@@ -216,7 +216,7 @@ But....if the push button is greyed out go to the section [Greyed out push butto
 
 1. Pull will take any changes to the global repo and bring them into your local repo. Go to your example GitHub repo (online) and click on your test file > edit this file.
 
-1. Add a line of code or a comment, preview the changes then commit directly to the master branch.
+1. Add a line of code or a comment, preview the changes then commit directly to the main branch.
 
 ```{r echo=FALSE, out.width = "800pt", fig.align='center', cache=TRUE}
 knitr::include_graphics('prac4_images/pullexample.png') 
@@ -273,16 +273,17 @@ Then check the current branch in RStudio (and Git) is tracking a branch on the r
 git branch -vv
 
 ## output
-master 3abe929 [origin/main] test3
+main 3abe929 [origin/main] test3
 ```
 
-Origin/master shows that the local master is tracking the origin/master on the remote repo. If you can't see origin/master then set it using:
+Origin/main shows that the local main is tracking the origin/main on the remote repo. If you can't see origin/main then set it using the following code. At the moment RStudio and git still defaults to the starting branch of master so the fist line below will change it to main --- which is required to match with the remote (GitHub).
 
 ```{r, eval=FALSE, cache=TRUE}
-git push --set-upstream origin main
+git branch -M main
+git push -u origin main
 ```
 
->Origin is the repository you cloned (from GitHub) and master is the name of the branch. You might see something like...your branch is ahead of origin/master by 1 commit. This means you have commited something you are working on in you local repo (Git) that hasn't yet been pushed to GitHub (the origin) and master branch...GitHub defaults the first branch to be called master.
+>Origin is the repository you cloned (from GitHub) and main is the name of the branch. You might see something like...your branch is ahead of origin/main by 1 commit. This means you have commited something you are working on in you local repo (Git) that hasn't yet been pushed to GitHub (the origin) and main branch...GitHub defaults the first branch to be called main
 
 If you need to change the URL of your GitHub .... so where you push your local Git to the GitHub account (changing this), perhaps you have made a new GitHub repo...
 
@@ -363,11 +364,11 @@ knitr::include_graphics('prac4_images/branch.png')
 knitr::include_graphics('prac4_images/branchtest.png') 
 ```
 
[…]
