+++ 06-prac6.Rmd
@@ -255,31 +255,34 @@ Qcount <- BluePlaquesSub.ppp %>%
   dplyr::count(Var1=Freq)%>%
   dplyr::rename(Freqquadratcount=n)
 
-#we don't need the last row, so remove it
-QCountTable <- Qcount %>%
-  slice(.,
-        -(nrow(Qcount)))
 ```
 
 Check the data type in the first column --- if it is factor, we will need to convert it to numeric
 
 ```{r, cache=TRUE}
-QCountTable %>% 
+Qcount %>% 
   summarise_all(class)
 ```
 
 OK, so we now have a frequency table --- next we need to calculate our expected values. The formula for calculating expected probabilities based on the Poisson distribution is:
 
 $$Pr= (X =k) = \frac{\lambda^{k}e^{-\lambda}}{k!}$$
+where:
+
+* `x` is the number of occurrences 
+
+* `λ` is the mean number of occurrences
+
+* `e` is a constant- 2.718
 
 ```{r, cache=TRUE}
-sums <- QCountTable %>%
+sums <- Qcount %>%
   #calculate the total blue plaques (Var * Freq)
   mutate(total = Var1 * Freqquadratcount) %>%
   dplyr::summarise(across(everything(), sum))%>%
   dplyr::select(-Var1) 
 
-lambda<- QCountTable%>%
+lambda<- Qcount%>%
   #calculate lambda
   mutate(total = Var1 * Freqquadratcount)%>%
   dplyr::summarise(across(everything(), sum)) %>%
@@ -289,13 +292,15 @@ lambda<- QCountTable%>%
 
 ```
 
+
 Calculate expected using the Poisson formula from above $k$ is the number of blue plaques counted in a square and is found in the first column of our table...
 
 ```{r, cache=TRUE}
 
-QCountTable <- QCountTable %>%
+QCountTable <- Qcount %>%
   mutate(Pr=((lambda^Var1)*exp(-lambda))/factorial(Var1))%>%
-  #now calculate the expected counts and save them to the table
+  #now calculate the expected counts based on our total number of plaques
+  #and save them to the table
   mutate(Expected= (round(Pr * sums$Freqquadratcount, 0)))
 
 #Compare the frequency distributions of the observed and expected point patterns
@@ -306,14 +311,17 @@ points(QCountTable$Freqquadratcount,
        col="Red", 
        type="o", 
        lwd=3)
-points(QCountTable$Expected, col="Blue", 
+points(Qcount$Expected, col="Blue", 
        type="o", 
        lwd=3)
 ```
+Comparing the observed and expected frequencies for our quadrat counts, we can observe that they both have higher frequency counts at the lower end --- something reminiscent of a Poisson distribution. This could indicate that for this particular set of quadrants, our pattern is close to Complete Spatial Randomness (i.e. no clustering or dispersal of points). But how do we confirm this?
 
-Comparing the observed and expected frequencies for our quadrat counts, we can observe that they both have higher frequency counts at the lower end --- something reminiscent of a Poisson distribution. This could indicate that for this particular set of quadrats, our pattern is close to Complete Spatial Randomness (i.e. no clustering or dispersal of points). But how do we confirm this?
+To check for sure, we can use the `quadrat.test()` function, built into `spatstat`. This uses a Chi Squared test to compare the observed and expected frequencies for each quadrant (rather than for quadrant bins, as we have just computed above).
 
[…]
