### 2020-11-18 539d62f
Update 06-prac6.Rmd
