### 2020-11-19 868ed3a
Update 06-prac6.Rmd
less noisy convex hull polygons in task 6.7
