### 2020-11-19 39d3f4c
Update 06-prac6.Rmd
