+++ 06-prac6.Rmd
@@ -535,17 +535,33 @@ Before we get any further, let's get some ward boundaries read in to R --- downl
 library(here)
 #read the ward data in
 LondonWards <- st_read(here::here("prac6_data", "LondonWards.shp"))
+
+LondonWardsMerged <- st_read(here::here("prac6_data", 
+                                        "statistical-gis-boundaries-london", 
+                                        "statistical-gis-boundaries-london", 
+                                        "ESRI",
+                                        "London_Ward_CityMerged.shp"))%>%
+  st_transform(.,27700)
+WardData <- read_csv("https://data.london.gov.uk/download/ward-profiles-and-atlas/772d2d64-e8c6-46cb-86f9-e52b4c7851bc/ward-profiles-excel-version.csv", 
+                     na = c("NA", "n/a")) %>% 
+  clean_names()
+
+LondonWardsMerged <- LondonWardsMerged %>% 
+  left_join(WardData, 
+            by = c("GSS_CODE" = "new_code"))%>%
+  distinct(GSS_CODE, ward_name, average_gcse_capped_point_scores_2014)
+
 ```
 
-It's probably projected correctly, but in case it isn't give it a projection using the ```CRS()``` function in the ```raster``` package
+It's probably projected correctly, but in case it isn't give it a projection using the ```st_crs()``` function in the ```sf``` package
 
 ```{r, cache=FALSE, message=FALSE, warning=FALSE}
 #have a look to check that it's 
 #in the right projection
-st_crs(LondonWards)
+st_crs(LondonWardsMerged)
 
 tmap_mode("view")
-tm_shape(LondonWards) +
+tm_shape(LondonWardsMerged) +
   tm_polygons(col = NA, alpha = 0.5) +
 tm_shape(BluePlaques) +
   tm_dots(col = "blue")
@@ -557,9 +573,9 @@ Ah yes, we might need to lose the blue plaques that fall outside of London
 
 ```{r, cache=FALSE}
 summary(BluePlaques)
-BluePlaquesSub <- BluePlaques[LondonWards,]
+BluePlaquesSub <- BluePlaques[LondonWardsMerged,]
 
-tm_shape(LondonWards) +
+tm_shape(LondonWardsMerged) +
   tm_polygons(col = NA, alpha = 0.5) +
 tm_shape(BluePlaquesSub) +
   tm_dots(col = "blue")
@@ -573,15 +589,16 @@ To create a continuous observation from the blue plaques data we need to count a
 
 ```{r echo=TRUE, results='hide', cache=FALSE}
 library(sf)
-points_sf_joined <- LondonWards%>%
+points_sf_joined <- LondonWardsMerged%>%
   st_join(BluePlaquesSub)%>%
-  add_count(WardName)%>%
-  #calcualte area
+  add_count(ward_name)%>%
+  janitor::clean_names()%>%
+  #calculate area
   mutate(area=st_area(.))%>%
   #then density of the points per ward
   mutate(density=n/area)%>%
   #select density and some other variables 
-  dplyr::select(density, WardName, Wardcode, n, AvgGCSE201)
+  dplyr::select(density, ward_name, gss_code, n, average_gcse_capped_point_scores_2014)
 
 ```
 
@@ -590,9 +607,9 @@ How about a quick choropleth map to see how we are getting on...
 ```{r, cache=FALSE}
 
 points_sf_joined<- points_sf_joined %>%                    
-  group_by(Wardcode) %>%         
+  group_by(gss_code) %>%         
   summarise(density = first(density),
-          wardname= first(WardName),
[…]
