### 2020-11-25 3a6f21f
Update 07-prac7.Rmd
upper case
