+++ 08-prac8.Rmd
@@ -62,7 +62,6 @@ library(ggplot2)
 library(reshape2)
 library(plotly)
 library(raster)
-library(downloader)
 library(rgdal)
 ```
 
