### 2020-12-09 03176aa
added more info on spatially lagged regression
