### 2020-12-09 a18d3e1
Update 09-prac9.Rmd
resetting plot layout
