+++ 10-prac10.Rmd
@@ -708,7 +708,7 @@ min_max_accuracy <- mean(apply(actuals_preds, 1, min, na.rm=TRUE) /
 min_max_accuracy
 ```
 
-22. This gives around a 60% accuracy...another alternative is mean absolute percentage deviation which is a statistical measure showing how accurate the prediciton was...through the equation...
+1. This gives around a 60% accuracy...another alternative is mean absolute percentage deviation which is a statistical measure showing how accurate the prediciton was...through the equation...
 
 $$MAPE= mean\left( \frac{abs(predicteds - actuals}{actuals} \right)$$
 
@@ -731,7 +731,7 @@ Here we've got a very high MAPE --- this is most likely because of the range of 
 knitr::include_graphics('prac10_images/cross_validation.jpg') 
 ```
 
-24. To start with we'll remove the geomtery, set up a data frame and replace any NAs with 0 --- this is appropraite here as if there are no Airbnbs within the ward they would be 0.
+1. To start with we'll remove the geomtery, set up a data frame and replace any NAs with 0 --- this is appropraite here as if there are no Airbnbs within the ward they would be 0.
 
 ```{r, cache=FALSE, warning=FALSE, message=FALSE}
 joined_CV <- joined %>%
@@ -853,7 +853,7 @@ joined3 <-joined2 %>%
             by= c("GSS_CODE.x" = "new_code"))
 ```
 
-28. Now run the model again
+1. Now run the model again
 
 ```{r, cache=FALSE, warning=FALSE, message=FALSE}
 model3<- train(Wardcount ~ Boroughcount+population_2015
@@ -878,36 +878,36 @@ Our R squared and adjusted R squared have increased (albeit slightly) and all pr
 knitr::include_graphics('prac10_images/ridge.jpg') 
 ```
 
-Least squares regression tries to minimise how far each point is away from the line...but if you look in this figure...the red dots are the training data and the green dots are the test data. Here normal OLS has been fitted to the training data perfcetly...the sum of the sqaured residuals (squared error distance from the line) is very small, probably 0. But when applied to the test data the sum of the squared residuals is large --- it therefore has high variance and would be termed to be overfitting the training data.
+Least squares regression tries to minimise how far each point is away from the line...but if you look in this figure...the red dots are the training data and the green dots are the test data. Here normal OLS has been fitted to the training data perfectly...the sum of the squared residuals (squared error distance from the line) is very small, probably 0. But when applied to the test data the sum of the squared residuals is large --- it therefore has high variance and would be termed to be overfitting the training data.
 
 ```{r echo=FALSE, out.width = "300", fig.align='center',fig.cap="This is a screenshot from the YouTube video Regularization Part 1: Ridge Regression by StatQuest", cache=FALSE}
 knitr::include_graphics('prac10_images/ridge_eg.png') 
 ```
 
-Ridge regression probably won't fit our training data as well, becuase we introduce a small amount of bias (adding value to the intercept --- which is the Y value when X=0) to give us less overall variance (the difference between the point and line value)...So essentially when the values are calculated the formula is trying to get the line of best fit through all the points with each point having minimum distance to the line.
+Ridge regression probably won't fit our training data as well, because we introduce a small amount of bias (adding value to the intercept --- which is the Y value when X=0, normally making the line of best fit more horizontal) to give us less overall variance (the difference between the point and line value)...So essentially when the values are calculated the formula is trying to get the line of best fit through all the points with each point having minimum distance to the line. This is usually useful in trying to get predicitions from the model.
 
[…]
