+++ 10-prac10.Rmd
@@ -46,7 +46,7 @@ Some of these practicals are long, take regular breaks and have a listen to some
 
 ## Introduction
 
-The Modifiable Areal Unit Problem (MAUP) represents the related effects of scale and aggregation on all geographic data and analysis. It was first deomstrated by the geographer Stan Openshaw in 1984 who showed that as you aggregated results to diffrent spatial units the reults could be manipulated to show different outcomes. Throughout this practical book we've considered London boroughs, but would any of the results change if we considered the data at ward level? Are we hiding any trends because we have just used the borough data that summed all of the wards within the borough? It's important to consider what is the most appropraite spatial unit for your analysis and provide appropraite reasoning. Even this is pretty straightforward (e.g. the data was only provided at the borough level) you must contextualise this and describe any potential limitations of it. In this practical i will firstly demonstrate the MAUP in action using some more advanced R code. Then we will have a look at some technqieus to model and validate data.
+The Modifiable Areal Unit Problem (MAUP) represents the related effects of scale and aggregation on all geographic data and analysis. It was first deomstrated by the geographer Stan Openshaw in 1984 who showed that as you aggregated results to diffrent spatial units the reults could be manipulated to show different outcomes. Throughout this practical book we've considered London boroughs, but would any of the results change if we considered the data at ward level? Are we hiding any trends because we have just used the borough data that summed all of the wards within the borough? It's important to consider what is the most appropriate spatial unit for your analysis and provide appropriate reasoning. Even this is pretty straightforward (e.g. the data was only provided at the borough level) you must contextualise this and describe any potential limitations of it. In this practical i will firstly demonstrate the MAUP in action using some more advanced R code. Then we will have a look at some techniques to model and validate data.
 
 Here we will be using London brough and ward data from practical 1. As we're getting better with R, we will try to automate almost everthing --- meaning that if you gave this code to someone else they could just run it without any data files and generate the same result. The only thing we won't automate later on is loading an excel file... i did find a function online that would let us read it from the interweb but it was more hassle than it was worth. I don't know why the data isn't just also distributed as a `.csv`.
 
@@ -74,7 +74,7 @@ listfiles<-dir_info(here::here("prac10_data")) %>%
   utils::unzip(exdir=here::here("prac10_data"))
 ```
 
-Using a recent update we could also change this a bit with the `pins()` package that downlaods and caches the url and automatically checks for file changes, only re-downloading if needed!
+Using a recent update we could also change this a bit with the `pins()` package that downloads and caches the url and automatically checks for file changes, only re-downloading if needed!
 
 ```{r, cache=FALSE, warning=FALSE, message=FALSE, results=FALSE}
 library(pins)
@@ -83,7 +83,7 @@ pinexample<-pin("https://data.london.gov.uk/download/statistical-gis-boundary-fi
 
 `pins()` also lets you share data easily, [have a look the RStudio pins blog post](https://blog.rstudio.com/2019/09/09/pin-discover-and-share-resources/) for more information.
 
-1. Take the downloaded data and filter it based on the filename that countains: Borough OR Ward_ AND `.shp` using `grepl()`
+1. Take the downloaded data and filter it based on the filename that contains: Borough OR Ward_ AND `.shp` using `grepl()`
 
 ```{r, cache=FALSE, warning=FALSE, message=FALSE}
[…]
