### 2021-01-04 #85 @andrewmaclachlan
Hi @labarba and @arfon

Just wondering if this has been accepted as i've found it here under accepted papers here: https://github.com/openjournals/jose-papers/blob/jose.00085/jose.00085/10.21105.jose.00085.pdf

Please let me know if i need to do anything.

Thanks
