+++ packages.bib
@@ -3,32 +3,32 @@
   author = {{R Core Team}},
   organization = {R Foundation for Statistical Computing},
   address = {Vienna, Austria},
-  year = {2020},
+  year = {2021},
   url = {https://www.R-project.org/},
 }
 
 @Manual{R-bookdown,
   title = {bookdown: Authoring Books and Technical Documents with R Markdown},
   author = {Yihui Xie},
-  year = {2020},
-  note = {R package version 0.20.2},
-  url = {https://github.com/rstudio/bookdown},
+  year = {2021},
+  note = {R package version 0.22},
+  url = {https://CRAN.R-project.org/package=bookdown},
 }
 
 @Manual{R-knitr,
   title = {knitr: A General-Purpose Package for Dynamic Report Generation in R},
   author = {Yihui Xie},
-  year = {2020},
-  note = {R package version 1.29},
-  url = {https://CRAN.R-project.org/package=knitr},
+  year = {2021},
+  note = {R package version 1.32},
+  url = {https://yihui.org/knitr/},
 }
 
 @Manual{R-rmarkdown,
   title = {rmarkdown: Dynamic Documents for R},
   author = {JJ Allaire and Yihui Xie and Jonathan McPherson and Javier Luraschi and Kevin Ushey and Aron Atkins and Hadley Wickham and Joe Cheng and Winston Chang and Richard Iannone},
-  year = {2020},
-  note = {R package version 2.4.6},
-  url = {https://github.com/rstudio/rmarkdown},
+  year = {2021},
+  note = {R package version 2.7},
+  url = {https://CRAN.R-project.org/package=rmarkdown},
 }
 
 @Book{bookdown2016,
@@ -38,7 +38,7 @@
   address = {Boca Raton, Florida},
   year = {2016},
   note = {ISBN 978-1138700109},
-  url = {https://github.com/rstudio/bookdown},
+  url = {https://bookdown.org/yihui/bookdown},
 }
 
 @Book{knitr2015,
