+++ docs/reference-keys.txt
@@ -1,0 +1,351 @@
+fig:unnamed-chunk-139
+fig:unnamed-chunk-141
+fig:unnamed-chunk-147
+fig:unnamed-chunk-148
+fig:vectorplots
+fig:unnamed-chunk-524
+fig:unnamed-chunk-525
+geographic-information
+learning-outcomes
+homework
+recommended-listening
+the-basics-of-geographic-information
+data-types-in-statistics
+important-gis-data-formats
+shapefiles
+geojson
+shapefile-and-geojson
+raster-data
+other-data-formats
+geodatabase
+geopackage
+spatialite
+postgis
+what-will-i-use
+general-data-flow
+data
+file-paths
+data-download
+data-pre-processing
+qgis-1
+load-data
+join-data
+export-data
+r-1
+introduction
+load-data-1
+join-data-1
+export-data-1
+making-a-script
+what-will-i-use-1
+data-sources-and-task
+uk-data-service
+ons
+nomis
+os
+edina-digimap
+osm
+defra
+data-lists
+summary
+feedback
+mac-r-issues
+introduction-to-r
+learning-outcomes-1
+homework-1
+recommended-listening-1
+introduction-1
+online-forums-are-your-friend
+health-warning
+r-and-rstudio
+getting-started-1
+basics
+scripts-and-some-basic-commands
+functions
+basic-plotting
+help
+data-structures
+elements-of-a-data-frame
+reading-data-into-r
+old-skool-cleaning
+here
+new-skool-cleaning
+examining-your-new-data
+data-manipulation-in-r
+selecting-rows
+selecting-columns
+renaming-columns
+more-dplyr-verbs
+levelling-up-withdplyr
+plotting
+pimp-my-graph
+spatial-data-in-r
+background-to-spatial-data-in-r
+making-some-choropleth-maps
+attribute-data
+simple-mapping
+tidying-data
+feedback-1
+rasters-descriptive-statistics-and-interpolation
+learning-outcomes-2
+homework-2
+recommended-listening-2
+introduction-2
+part-1-projections
+changing-projections
+vector
+proj4
+epsg
+reprojecting-your-spatial-data
+worldclim-data
+worldclim-data-loading-efficiently
+location-data-from-a-raster
+part-2-descriptive-statistics
+data-preparation
+histogram
+using-more-data
+histogram-with-ggplot
+part-3-interpolation
+auto-data-download
+advanced-analysis
+feedback-2
+git-github-and-rmarkdown
+learning-outcomes-3
+homework-3
+recommended-listening-3
+introduction-3
+git-and-github
+the-three-ways
+set-up-your-github
+using-rstudio-with-git
+using-the-git-gui---way-1
+create-a-new-version-control-in-rstudio---way-2
+if-have-have-an-existing-project---way-3
+commiting-to-git
+using-git-outside-rstudio
+push-to-github
+pull-from-github
+troubleshooting
+were-you-challenged-for-your-password
+greyed-out-push-button
+reprex
+fork-a-repository
+branches
+back-in-time
+git
+github
+on-final-trick
+git-commands
+health-warning-1
+rmarkdown
+html
+flexdashboard
+word-document
+knit-options
+shortcuts
+adding-references
+set-up-mendeley
+add-refereces-into-r
+references-using-citr
+yaml-options
+packrat
+binder
+further-reading
+feedback-3
+map-making
+learning-outcomes-4
+homework-4
+recommended-listening-4
+introduction-4
+osm-1
+airbnb
+data-1
+osm-2
+london-boroughs
+world-cities
+airbnb-1
+qgis-2
+load-data-2
+manipulate-data
+map-data
+export-map
+graphical-modeler
+r-2
+static-map
[…]
