### 2021-05-24 #85 @labarba
@darribas 👋  — You [published](https://jose.theoj.org/papers/10.21105/jose.00042) with JOSE a while back. Could I persuade you to help us out with a review? We have received this submission that I think is right up your alley: "An Applied Geographic Information Systems and Science Course in R" https://github.com/openjournals/jose-reviews/issues/85
Let me know…
