### 2021-05-24 #85 @labarba
@jsta 👋  — You [reviewed](https://jose.theoj.org/papers/10.21105/jose.00042) for JOSE a while back. Would you be able to contribute a review once more? We have a new submission that I think you could help us with: "An Applied Geographic Information Systems and Science Course in R" https://github.com/openjournals/jose-reviews/issues/85
Let me know…
