+++ docs/reference-keys.txt
@@ -3,8 +3,6 @@ fig:unnamed-chunk-141
 fig:unnamed-chunk-147
 fig:unnamed-chunk-148
 fig:vectorplots
-fig:unnamed-chunk-524
-fig:unnamed-chunk-525
 geographic-information
 learning-outcomes
 homework
@@ -105,8 +103,6 @@ data-preparation
 histogram
 using-more-data
 histogram-with-ggplot
-part-3-interpolation
-auto-data-download
 advanced-analysis
 feedback-2
 git-github-and-rmarkdown
@@ -128,28 +124,20 @@ pull-from-github
 troubleshooting
 were-you-challenged-for-your-password
 greyed-out-push-button
-reprex
 fork-a-repository
 branches
 back-in-time
 git
 github
-on-final-trick
+one-final-trick
 git-commands
 health-warning-1
 rmarkdown
 html
-flexdashboard
-word-document
 knit-options
 shortcuts
-adding-references
-set-up-mendeley
-add-refereces-into-r
-references-using-citr
-yaml-options
-packrat
-binder
+renv
+referencing-and-more
 further-reading
 feedback-3
 map-making
@@ -177,7 +165,7 @@ basic-interactive-map
 advanced-interactive-map
 bad-maps
 feedback-4
-analysing-spatial-patterns
+detecting-spatial-patterns
 learning-outcomes-5
 homework-5
 recommended-listening-5
@@ -202,11 +190,48 @@ mapping-outputs
 other-variables
 extension
 feedback-5
-advanced-raster-analysis
+explaining-spatial-patterns
 learning-objectives
 homework-6
 recommended-listening-6
 introduction-6
+setting-up-your-data-1
+cleaning-the-data-as-you-read-it-in
+additional-data
+analysing-gcse-exam-performance---testing-a-research-hypothesis
+research-question-and-hypothesis
+regression-basics
+running-a-regression-model-in-r
+interpreting-and-using-the-model-outputs
+broom
+bootstrap-resampling
+assumptions-underpinning-linear-regression
+assumption-1---there-is-a-linear-relationship-between-the-dependent-and-independent-variables
+transforming-variables
+should-i-transform-my-variables
+assumption-2---the-residuals-in-your-model-should-be-normally-distributed
+assumption-3---no-multicolinearity-in-the-independent-variables
+variance-inflation-factor-vif
+assumption-4---homoscedasticity
+assumption-5---independence-of-errors
+standard-autocorrelation
+spatial-regression-models
+dealing-with-spatially-autocorrelated-residuals---spatial-lag-and-spatial-error-models
+the-spatial-lag-lagged-dependent-variable-model
+the-spatial-error-model
+key-advice
+which-model-to-use
+more-data
+extending-your-regression-model---dummy-variables
+task-investigating-further---adding-more-explanatory-variables-into-a-multiple-regression-model
+a-few-things-to-watch-out-for
+task-3---spatial-non-stationarity-and-geographically-weighted-regression-models-gwr
+feedback-6
+advanced-raster-analysis
+learning-objectives-1
+homework-7
+recommended-listening-7
+introduction-7
 gitignore
 remote-sensing-background-required
 data-2
@@ -233,101 +258,7 @@ considerations
 extension-1
 references
 remote-sensing-background-optional
-feedback-6
-online-mapping-descriptive-statistics
-homework-7
-recommended-listening-7
-part-a-spatial-descriptive-statistics
-learning-objectives-1
-getting-started-2
-main-tasks
-task-1---descriptive-statistics
-extension-1-1
-task-2---function-to-recode-data
-whats-going-on-in-this-function
-task-3-mapping-location-quotients
-task-4-creating-a-basic-geodemographic-classification
-part-b-online-mapping
[…]
