+++ _08-prac8.Rmd
@@ -1,0 +1,403 @@
+# Descriptive statistics 
+
+# Learning objectives
+
+By the end of this practical you should be able to:
+
+1. Create descriptive plots (histograms and boxplots) to help understand the frequency distributions of your data
+1. Write custom functions to process your data
+1. Produce a location quotient map to highlight interesting (above and below average) patterns in your data
+1. Write a function in R to produce a range of different maps based on user inputs
+1. Perform a very basic cluster analysis and output the results of a basic geodemographic classification
+
+# Homework
+
+Outside of our schedulded sessions you should be doing around 12 hours of extra study per week. Feel free to follow your own GIS interests, but good places to start include the following:
+
+::: {.infobox .assignment data-latex="{note}"}
+
+**Assignment**
+
+From weeks 6-9, learn and practice analysis from the course and identify appropriate techniques (from wider research) that might be applicable/relevant to your data. Conduct an extensive methodological review – this could include analysis from within academic literature and/or government departments (or any reputable source).
+:::
+
+::: {.infobox .note data-latex="{note}"}
+**Reading**
+
+* For k-means clustering and exploratory data analysis read [Chapter 12 "K-Means Clustering"](https://bookdown.org/rdpeng/exdata/k-means-clustering.html) from Exploratory Data Analysis with R by Peng (2016).
+
+Remember this is just a starting point, explore the [reading list](https://rl.talis.com/3/ucl/lists/139FBAF8-DACD-60FB-8BDC-E9C9E09BA885.html?lang=en-GB&login=1), practical and lecture for more ideas.
+:::
+
+## Recommended listening `r emo::ji("headphones")`
+
+Some of these practicals are long, take regular breaks and have a listen to some of our fav tunes each week.
+
+[Andy](https://www.youtube.com/watch?v=6Bh6IL1mNfc)
+
+
+[Adam](https://open.spotify.com/track/3BGF4XAm8jaEi2LcGR57O3?si=gfYc6LS6Tgy9M9oXVVG4SQ)
+Bit different from me this week - I'm not all about drum and bass (although I mostly am). Stumbled across these recently and I should have known about them for much longer. The Joy Formidable - this is a 10th Anniversary revisit of their first EP and is ace. If you like Welsh, you're in for a treat! 
+
+
+# Getting Started
+Before we begin this week’s practical, we need to load our packages and carry out some data preparation... 
+
+```{r prac8_load, message=FALSE}
+library(highcharter)
+library(tidyverse)
+library(downloader)
+library(rgdal)
+library(sf)
+library(ggplot2)
+library(reshape2)
+library(plotly)
+library(raster)
+library(downloader)
+library(rgdal)
+```
+
+
+There is a problem with our London Wards data --- we are missing some data relating to housing tenure. The housing tenure data in this file comes from the 2011 Census and visiting http://www.nomisweb.co.uk/ and interrogating Table KS402EW (the Tenure table), we can discover that data for the percentage of shared owners and those living in accommodation rent free are missing. 
+
+Rather than making you go off to Nomisweb and fetch this data, because I'm really nice, I've posted on GitHub a file containing this and extra categorical, ratio and geographical data that we will need to add to our existing London data file. To download this consult [How to download data and files from GitHub], i'd used Option 1 and it's the prac8_data folder you want to download.
+
+We can easily join this new data to our original data in R.
+
+```{r prac8_read_data}
+
[…]
