### 2021-07-27 3e7c948
change some GWR
