+++ 07-prac7.Rmd
@@ -346,186 +346,7 @@ Regressiondata %>%
 
 ### Bootstrap resampling
 
-If we only fit our model once, how can we be confident about that estiamte? Bootstrap resampling is where we take the original dataset and select random data points from within it, but in order to keep it the same size as the original dataset some records are duplicated. This is known as bootstrap resampling by replacement.
-
-So first let's again get rid of all the coloumns we aren't using in this model using as they will just slow us down. At the moment we only care about average GCSE scores and unauthorised absecne in schools. 
-
-```{r, message=FALSE, cache=FALSE}
-Bootstrapdata<- LonWardProfiles%>%
-  clean_names()%>%
-  dplyr::select(average_gcse_capped_point_scores_2014, 
-                unauthorised_absence_in_all_schools_percent_2013)
-```
-
-Next we are going to create our samples using the function `bootstraps()`. First we set the seed (to any number), which just means are results will be the same if we are to re-run this analysis, if we don't then as we are randomly replacing data the next time we do run the code the results could change. 
-
-Here we have created 1000 versions of our original data (times=1000). `apparent` means we are using our entire dataset and the full dataset will be kept in our `GSCE_boot` variable. `set.seed(number)` just makes sure we get the same result here each time as it could vary.
-
-```{r, message=FALSE, cache=FALSE}
-library(rsample)
-set.seed(99)
-
-GCSE_boot <-st_drop_geometry(Bootstrapdata) %>%
-  bootstraps(times = 1000, apparent = TRUE)
-
-
-slice_tail(GCSE_boot, n=5)
-```
-So when we look in `GSCE_boot` we have 1000 rows, plus our `apparent` data.
-
-Here we are now going to run the linear regression model for each of our bootstrap resampled instances. To do this we are going to use the `map()` function from the `purrr` package that is also part of the `tidyverse`. This lets us apply a function (so our regression formula) to a list (so our splits that we calculated from bootstrap resampling)
-
-We need to map over and then given the function to what we are mapping over. So we are mapping (or changing) our splits (from the bootstrap resampled) then apply our regression formula to each split. To make sure we can save the output we will create a new column called model with the `mutate()` function, remember this just lets us add a column to an existing table.
-
-```{r, message=FALSE, cache=FALSE}
-
-GCSE_models <- GCSE_boot %>%
-  #make new column
-  mutate(
-    #column name is model that contains...
-    model = map(splits, ~ lm(average_gcse_capped_point_scores_2014 ~ unauthorised_absence_in_all_schools_percent_2013, 
-                             data = .)))
-
-# let's look at the first model results
-GCSE_models$model[[1]]
-
-#if you wanted all of them it's
-#GCSE_models$model
-```
-
-However, this is a very ugly output and pretty hard to understand. Now remember when we just went through the magic of [broom]? You might have thought that it didn't do very much! This is where it can really be of use. We want to tidy up our models results so we can actually use them. Firstly let's add a new coloumn with mutate (coef_info) then take our model coloumn and using the `map()` function again tidy the outputs (saving the result into that new coef_info column).
-
-```{r, message=FALSE, cache=FALSE}
-
-GCSE_models_tidy <- GCSE_models %>%
-  mutate(
-    coef_info = map(model, tidy))
-```
-
[…]
