### 2021-07-30 f33b356
practical 1 and 2 tasks
