+++ 02-prac2.Rmd
@@ -17,7 +17,15 @@ Outside of our schedulded sessions you should be doing around 12 hours of extra 
 ::: {.infobox .assignment data-latex="{note}"}
 **Assignment**
 
-This week you need to assess the validity of your topic in terms of data suitability, then sketch out an introduction and literature review including academic and policy documents (or any reputable source).
+Each week we will provide a short task to test your knowledge, these should be used to guide your study for the final exam.
+
+The task is to join some non spatial data to some spatial data and wrangle it.
+
+You need create identify the average percent of science students (in all) grades per county meeting the required standards and produce a map to show where the Country averages are above or below the State of Washington average.
+
+* [Download Washing Counties](https://hub.arcgis.com/datasets/kingcounty::washington-counties-with-natural-shoreline-washsh-area/explore?location=47.272700%2C-120.740400%2C7.67)
+
+* [Download school report card data](https://data.wa.gov/Education/Report-Card-Assessment-Data-2018-19-School-Year/5y3z-mgxd)
 :::
 
 ::: {.infobox .note data-latex="{note}"}
