### 2021-08-19 75a7842
update weekly tasks
