+++ 12-assignment_resources.Rmd
@@ -1,32 +1,26 @@
-# (PART\*) Assignment {-}
+# (PART\*) Extra {-}
 
-# Assignment resources
+# Module resources
 
 ```{r echo=FALSE}
 library(knitr)
 ```
 
-Listed below are some helpful website, books, datasets and tools you might find useful when completing your assignments.
-
-## Timeline
-
-There is a suggested timeline of work for your assignment provided in the GIS coursework requirements PDF on Moodle. 
+Listed below are some helpful website, books, datasets and tools you might find useful.
 
 ## Markscheme
 
-The markscheme can be found on Moodle with a break down of how much percent is given to each criteron. Read it closely **and often** when writing your assignment.
-
-## Examples of previous work
-
-Closer to the assignment deadline we will release examples of good student projects from previous years as a tab on Moodle. **Note** that the requirements / markscheme have been altered this year.
+The markscheme can be found on Moodle with a break down of how much percent is given to each criteron. Read it closely **and often** when preparing for the final exam.
 
 ## Events / meet other spatial data professionals
 
-The [UK ESRI conference](https://www.esriuk.com/en-gb/about/events/ac/overview) is normally free to go to and whilst (in my opinion) it's mostly just ESRI marketing their latest products it is good to go along and see how industry are using spatial software. It might also be useful for dissertation ideas / networking. Next year it's scheduled for the 19/5/2020.
+The [UK ESRI conference](https://www.esriuk.com/en-gb/about/events/ac/overview) is normally free to go to and whilst (in my opinion) it's mostly just ESRI marketing their latest products it is good to go along and see how industry are using spatial software. It might also be useful for dissertation ideas / networking. 
 
 [Missing maps](http://www.missingmaps.org/events/)
 
 [LondonR group](https://www.londonr.org/#content)
+
+[R Ladies](https://www.meetup.com/pro/rladies/)
 
 ## Cool stuff to explore
 
@@ -40,13 +34,6 @@ New packages and functions that i've recently come across that are worth explori
 * [R Cartography package](http://riatelab.github.io/cartography/docs/)
 * [Data is beautiful reddit](https://www.reddit.com/r/dataisbeautiful/)
 * [Network visualisation](https://kateto.net/network-visualization)
-</div><br/>
-
-## Assignment ideas
-<div class="col2">
-
-* [The Value of GIS & Spatial Data Science in Achieving the Sustainable Development Goals by Cobus van Rooyen](https://complexsystemstheory.net/2020/09/23/gis-and-the-sustainable-development-goals/)
-
 </div><br/>
 
 ## Books/reading resources
@@ -181,76 +168,4 @@ This is by no means an extensive data list, but summarises data used within some
 ## Data lists
 
 [Awesome public datasets](https://github.com/awesomedata/awesome-public-datasets) have a wide range all data (some geographic, some not): 
-
 [Robin Wilson](https://twitter.com/sciremotesense) has authored one of [the most extensive data lists that i've come across](https://freegisdata.rtwilson.com/)
-
-## Reading and writing critically
-
-[Hugh Kerans](https://twitter.com/ithinkwellhugh?lang=en) from Flinders Unversity, Adelaide has developed a range of materials to assist graduate students in producing effective reports and publications...check out his website and resouces here: https://www.ithinkwell.com.au/
-
-The most useful templates for the assignments are the critical reading one:
-
-```{r echo=FALSE, out.height = "600pt", out.width = "800pt", cache=TRUE}
-include_graphics('assignment/CriticalReading.pdf', auto_pdf = getOption("knitr.graphics.auto_pdf", FALSE)
-)
-```
-
[…]
