+++ docs/reference-keys.txt
@@ -84,7 +84,7 @@ attribute-data
 simple-mapping
 tidying-data
 feedback-1
-descriptive-statistics
+spatial-descriptive-statistics
 learning-outcomes-2
 homework-2
 recommended-listening-2
@@ -259,24 +259,12 @@ extension-1
 references
 remote-sensing-background-optional
 feedback-7
-assignment-resources-1
-timeline
+module-resources
 markscheme
-examples-of-previous-work
 events-meet-other-spatial-data-professionals
 cool-stuff-to-explore
-assignment-ideas
 booksreading-resources
 new-developments
 twitter
 data-3
 data-lists-1
-reading-and-writing-critically
-free-flow-diagram-tools
-basics-1
-introduction-8
-literature-review
-methodology
-results
-discussion
-conclusion
