+++ 07-prac7.Rmd
@@ -12,11 +12,20 @@ By the end of this practical you should be able to:
 ## Homework 
 
 ::: {.infobox .assignment data-latex="{note}"}
-Outside of our schedulded sessions you should be doing around 12 hours of extra study per week. Feel free to follow your own GIS interests, but good places to start include the following:
 
-**Assignment**
+Outside of our scheduled sessions you should be doing around 12 hours of extra study per week. Feel free to follow your own GIS interests, but good places to start include the following:
 
-From weeks 6-9, learn and practice analysis from the course and identify appropriate techniques (from wider research) that might be applicable/relevant to your data. Conduct an extensive methodological review – this could include analysis from within academic literature and/or government departments (or any reputable source).
+**Exam**
+
+Each week we will provide a short task to test your knowledge, these should be used to guide your study for the final exam.
+
+For the task this week have a go at the practice exam questions, you should be able to answer either of the following questions:
+
+* New York Eviction
+* Graffiti mitigation
+
+The actual exam will be distributed using the same format, so it's vital to do this.
+
 :::
 
 ::: {.infobox .note data-latex="{note}"}
@@ -251,7 +260,7 @@ As a general rule, the better the blue line is at summarising the relationship b
 
 The equation for the blue line in the graph above can be written:
 
-(1)$$y_i = \beta_0 + \beta_1x_i + \epsilon_i$$
+$$y_i = \beta_0 + \beta_1x_i + \epsilon_i$$
 
 where:
 
@@ -346,7 +355,7 @@ Regressiondata %>%
 
 ### Bootstrap resampling
 
-If we only fit our model once, how can we be confident about that estimate? Bootstrap re sampling is where we take the original dataset and select random data points from within it, but in order to keep it the same size as the original dataset some records are duplicated. This is known as bootstrap resampling by replacement. We used to briefly cover this within this practical but have recently removed it. If you wish to explore it then consult the [bootstrap resampling section from previous years](https://andrewmaclachlan.github.io/CASA0005repo_20202021/gwr-and-spatially-lagged-regression.html#bootstrap-resampling), but this is not a requirement and only for interest. 
+If we only fit our model once, how can we be confident about that estimate? Bootstrap resampling is where we take the original dataset and select random data points from within it, but in order to keep it the same size as the original dataset some records are duplicated. This is known as bootstrap resampling by replacement. We used to briefly cover this within this practical but have recently removed it. If you wish to explore it then consult the [bootstrap resampling section from previous years](https://andrewmaclachlan.github.io/CASA0005repo_20202021/gwr-and-spatially-lagged-regression.html#bootstrap-resampling), but this is not a requirement and only for interest. 
 
 ### Assumptions Underpinning Linear Regression
 
@@ -751,7 +760,7 @@ As such the average GCSE score in one ward could be related to that in another a
 
 Ward and Gleditsch (2008) describe this situation (where the value of our $y$ dependent variable - GCSE scores - may be influenced by neighbouring values) and suggest the way to deal with it is to incorporate a *spatially-lagged* version of this variable amongst the independent variables on the right-hand side of the equation. In this case, Equation 1 would be updated to look like this:
 
-(2)$$y_i = \beta_0 + \beta_1x_i + \rho w_i.y_i + \epsilon_i$$
+$$y_i = \beta_0 + \beta_1x_i + \rho w_i.y_i + \epsilon_i$$
 
 In this equation, $w$ is the spatial weights matrix you generated and $w_i$ is vector of all neighbouring areas (in our case, Wards) for any Ward $y_i$. 
 
[…]
