+++ 04-prac4.Rmd
@@ -19,8 +19,9 @@ Each week we will provide a short task to test your knowledge, these should be u
 The task this week is to:
 
 * Read in [global gender inequality data](http://hdr.undp.org/en/indicators/68606#) 
-* Join the global gender inequality index to spatial data, creating a new column of difference in inequality between 2010 and 2019
+* Join the global gender inequality index to [spatial data of the World](https://hub.arcgis.com/datasets/2b93b06dc0dc4e809d3c8db5cb96ba69_0/explore?location=-2.688200%2C0.000000%2C1.41), creating a new column of difference in inequality between 2010 and 2019
 * Share it with the World on GitHub
+* Add you repository URL to the circulated spreadsheet
 
 :::
 

[…]
