+++ 08-prac8.Rmd
@@ -4,13 +4,13 @@
 
 By the end of this practical you should be able to:
 
-1. Explain and execute appropraite pre-proessing steps of raster data
+1. Explain and execute appropriate pre-proessing steps of raster data
 1. Replicate published methodologies using raster data
 1. Design new R code to undertake further analysis 
 
 ## Homework
 
-Outside of our schedulded sessions you should be doing around 12 hours of extra study per week. Feel free to follow your own GIS interests, but good places to start include the following:
+Outside of our scheduled sessions you should be doing around 12 hours of extra study per week. Feel free to follow your own GIS interests, but good places to start include the following:
 
 ::: {.infobox .assignment data-latex="{note}"}
 **Exam**
@@ -244,8 +244,8 @@ raster::compareRaster(listlandsat$LC08_L1TP_203023_20190513_20190521_01_T1_B1,
 
 lsatmask <- listlandsat %>%
   # now crop our temp data to the extent
-  crop(.,manchester_boundary)%>%
-  mask(.,  manchester_boundary)
+  raster::crop(.,manchester_boundary)%>%
+  raster::mask(.,  manchester_boundary)
 ```
 
 1. If all we wanted to do was clip our data, we could now change our filenames in the raster stack and write the ```.TIFF ``` files out again...
@@ -793,14 +793,263 @@ As p<0.05 is shows that are variables are have a statistically significant corre
 
 If you want more information on statistics in R go and read [YaRrr! A Pirate's Guide to R](https://bookdown.org/ndphillips/YaRrr/), chapter 13 on hypothesis tests.
 
+## LSOA/MSOA stats
+
+This is all useful, but what if we wanted to present this analysis to local leaders in Manchester to show which areas experience the highest temperature (remember, with the limitation of this being one day!)
+
+Next, we will aggregate our raster data to LSOAs, taking a mean of the pixles in each LSOA.
+
+LSOA data: https://data-communities.opendata.arcgis.com/datasets/5e1c399d787e48c0902e5fe4fc1ccfe3
+
+MOSA data: https://data.cambridgeshireinsight.org.uk/dataset/output-areas/resource/0e5ac3b8-de71-4123-a334-0d1506a50288
+
+Let's prepare the new spatial data:
+```{r, warning=FALSE, message=FALSE}
+library(dplyr)
+
+# read in LSOA data
+UK_LSOA <- st_read(here::here("prac7_data", 
+                                          "Lower_Super_Output_Area_(LSOA)_IMD2019__(WGS84)-shp",
+                                          "5878ebcf-f091-4bde-8864-aa547dd45afd2020330-1-8g4usn.8evuq.shp"))
+
+# project it to match Manchester boundary
+UK_LSOA <- UK_LSOA %>%
+  st_transform(., 32630)
+
+# read in MSOA and project it
+MSOA <- st_read(here::here("prac7_data",
+                           "MSOA_EngWal_Dec_2011_Generalised_ClippedEW_0",
+                           "Middle_Layer_Super_Output_Areas_December_2011_Generalised_Clipped_Boundaries_in_England_and_Wales.shp")) %>%
+   st_transform(., 32630)
+
+#select only MSOA within boundary
+manchester_MSOA <- MSOA[manchester_boundary, , op=st_within]
+
+#select only LSOA that intersect MSOA
+manchester_LSOA <- UK_LSOA[manchester_MSOA,]
+```
+
[…]
