+++ _output.yml
@@ -4,6 +4,9 @@ bookdown::gitbook:
   css: assets/style.css
   highlight: tango
   config:
+    search:
+      engine: fuse # lunr is the default
+      options: null # can override, see: https://fusejs.io/api/options.html
     toc:
       collapse: section
       scroll_highlight: yes
