+++ _output.yml
@@ -6,7 +6,8 @@ bookdown::gitbook:
   config:
     search:
       engine: fuse # lunr is the default
-      options: includeMatches # can override, see: https://fusejs.io/api/options.html
+      options: 
+        includeScore: true
     toc:
       collapse: section
       scroll_highlight: yes
