+++ Extras.Rmd
@@ -1,405 +1,0 @@
-# Extra material 
-
-## Shiny
-
-Shiny is an R package that lets you build interactive web apps, host them online and embed them within RMarkdown documents... have a [look at some examples](https://shiny.rstudio.com/gallery/)
-
-To build a shiny you require three main *'items'* or blocks of code:
-
-(a) Code that specfies the user interface or ```ui```
-
-(b) Code that defines the server logic to plot variables (server function)
-
-(c) A call to the ```ShinyApp``` function
-
-These could either be within one large ```.R``` file or over several ```.R``` files.
-
-25. Start a new R Project > New or existing directory > Shiny Web Application 
-
-26. Within your new R Project folder make a new folder called data and copy in the data we have been using
-
-The main purpose of this part of the practical is to show you how to use Shiny so we won't make a Git repository. 
-
-27. Go File > New File > Shiny Web App
-
-Here you can select either a single file or multiple file format, originally Shiny needed multiple files to run however in the latest version you can have this all in one script, i've just used one script here.
-
-28. If you click Run App on either the ```app.R``` the example Shiny will load.
-
-Now let's make one with our data...
-
-#### Data preparation
-
-29. The first part is easy and is based on analysis we have already compelted...we need to produce a ```sf``` multipolygon layer containing the number of hotels and airbnbs per London borough through....
-
-```{r prac8_shiny, eval=TRUE, cache=FALSE, eval=FALSE}
-# load packages
-library(sf)
-library(tmap)
-library(leafpop)
-library(leaflet)
-library(tmaptools)
-library(tidyverse)
-library(plyr)
-library(classInt)
-
-# read in OSM
-OSM <- st_read("data/gis_osm_pois_a_free_1.shp")
-
-# read in Londonboroughs
-Londonborough <- st_read("data/London_Borough_Excluding_MHW.shp")
-
-# read in Airbnb
-Airbnb <- read_csv("data/listings.csv")
-
-# plot xy data
-Airbnb <- st_as_sf(Airbnb, coords = c("longitude", "latitude"), 
-                   crs = 4326)
-
-# reproject
-OSM <- st_transform(OSM, 27700)
-Airbnb <- st_transform(Airbnb, 27700)
-# we don't need to reproject Londonborough, but it 
-# doesn't have a CRS..you could also use set_crs
-# it needs to have one for the next step
-Londonborough<- st_transform(Londonborough, 27700)
-
-#select hotels only
-OSM <- OSM[OSM$fclass == 'hotel',]
-Airbnb <- Airbnb[Airbnb$room_type == 'Entire home/apt' &
-                     Airbnb$availability_365=='365',]
-
-# make a function for the join
-# functions are covered in practical 7
-# but see if you can work out what is going on
-# hint all you have to do is replace data1 and data2
-# with the data you want to use
-
-Joinfun <- function(data1, data2) {
-    # join OSM and London boroughs
-    joined <- st_join(data1, data2, join = st_within)
-    
-    # count the number of hotels per borough
-    countno <- as.data.frame(count(joined$GSS_CODE))
-    
-    # join the count back to the borough layer
-    counted <-left_join(data2, countno, by=c("GSS_CODE"="x"))
-    
-    return(counted)
-}
-
-# use the function for hotels
-Hotels <- Joinfun(OSM, Londonborough)
-
-# then for airbnb
-Airbnb <- Joinfun(Airbnb, Londonborough)
-
-# now try to arrange the plots with tmap
-breaks = c(0, 5, 12, 26, 57, 286) 
-
-#change the column name from freq for the legend
[…]
