### 2021-09-16 #85 @labarba
@andrewmaclachlan — You signed up to review for JOSE a while back. Would you be available to help us with this submission? "An Applied Geographic Information Systems and Science Course in R" https://github.com/openjournals/jose-reviews/issues/85
Let me know…
