### 2021-10-06 #141 @whedon
[REVIEW]: An Applied Geographic Information Systems and Science Course in R
**Submitting author:** @andrewmaclachlan (<a href="http://orcid.org/0000-0003-0356-1623">Andrew MacLachlan</a>)
**Repository:** <a href="https://github.com/andrewmaclachlan/CASA0005repo" target ="_blank">https://github.com/andrewmaclachlan/CASA0005repo</a>
**Version:** 3.0
**Reviewer:** @TomasBeuzen @darribas

Due to the challenges of the COVID-19 pandemic, JOSE is currently operating in a "reduced service mode".

- [x] **Version:** 3.0

- [x] **Version:** 3.0
