### 2021-10-06 #141 @ttimbers
@darribas , please carry out your review in this issue by updating the checklist below. If you cannot edit the checklist please:

- [ ] **Version:** Does the release version given match the repository release (2.0)?
