### 2021-10-17 #141 @TomasBeuzen
@ttimbers - I can't tick the check-boxes above, presumably because I didn't accept the invite in time and now the link has expired 😄 I'm keeping a copy of the checklist below for now, but can remove it once we figure out the issue.

@andrewmaclachlan - firstly, excellent work putting together this course. I can see that a lot of hard-work went into creating this content and I think it's an excellent resource for other in the field. I provide general comments below:

## Review Comments

### Content, Pedagogy, Instructional Design

- All excellent and I don't see anything in particular to change here. The content is comprehensive, well presented, and uses modern tools (e.g., `tidyverse`). Admittedly I've been using Python for all my geospatial data analysis over the last few years, so not 100% sure what's new with R spatial world, but all the common tools are here, `spatstat`, `sp`, `sf`, etc.
- I really enjoy the casual but informative style of teaching. I may be a little biased because it's quite similar to my own teaching style, but I find the conversational nature of the text, backed-up liberally with metaphors, examples, figures, call-outs, etc. to be an effective way of teaching what can be quite complex topics. Kudos.
- I was a little sad to see interpolation dropped from the most recent version of the material. It's something I use frequently in my day-to-day as a data scientist, e.g., kriging, areal interp from one set of polygons to another, etc. Not saying anything needs to be changed here, just writing down my thoughts 😄
- There are various spelling mistakes throughout the text (the most recent I saw was 8.10.1 Calucating tempearture from Landsat data). Again, this is completely understandable with such a large volume of material and I don't see any reason to need to go and correct every single mistake, just pointing it out as something to be aware of in future iterations.
- May be worth updating to the new pipe operator? `|>`

- The only thing missing is a reference to what the author's expertise is in the area, which is one of the review checklist items (*Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?*). Not sure how important that is to have though and it didn't bother me as the quality of the material speaks for itself.

There are a few improvements I'd like to see here:

- [ ] The README is the first place I'll often visit when looking at other's material. I think it needs to include at least a brief paragraph (1-2 sentences) about what the material covers and who the target audience is.
- [ ] I would like to see clear instructions for how to install the material locally and develop it. Imagine someone wanting to use it with minimal experience with GitHub/Bookdown/etc, e.g., 1. Clone this repository `git clone git@github.com:andrewmaclachlan/CASA0005repo.git`. 2. Install R and RStudio. 3. Install Bookdown.... Or more helpful would be a Docker image - I wouldn't require it to pass this review but having personally worked with bookdown to build course materials like this, it's a life-saver for your future self and others trying to use your material, so worth considering.
- [ ] There are no contributing guidelines, as far as I could see. I note "Feedback" forms in the material itself. But I think some guidelines in the README of main repo are needed, i.e. are you accepting PR's or issues? If so, how would you like them structured, or what kind of things can someone contribute?
- [ ] There's an MIT license for the repo. It gets a little grey when sharing a repo with both original code and content. I usually associate see an MIT license with software, but not necessarily written content, figures, etc. A CC license might be a more suitable option here? Or worth considering as an additional license that covers the content. Apologies if this is already covered somewhere and I missed it.
- [ ] Figures throughout the material are missing captions. In particular, I'm missing references to original sources. I note that you acknowledge these various sources in the welcome page, but as readers may not read that page, I think it's appropriate to add references to original figures when they are used in the text.

## Temporary Checklist

- [x] As the reviewer I confirm that I have read the JOSE conflict of interest policy and that there are no conflicts of interest for me to review this work.

- [x] I confirm that I read and will adhere to the JOSE code of conduct.

- [x] Repository: Is the source for this learning module available at the repository url?
- [ ] License: Does the repository contain a plain-text LICENSE file with the contents of a standard license? (OSI-approved for code, Creative Commons for content)
- [x] Version: Does the release version given match the repository release (2.0)?
- [x] Authorship: Has the submitting author (@andrewmaclachlan) made visible contributions to the module? Does the full list of authors seem appropriate and complete?

- [ ] A statement of need: Do the authors clearly state the need for this module and who the target audience is?
- [ ] Installation instructions: Is there a clearly stated list of dependencies?
- [ ] Usage: Does the documentation explain how someone would adopt the module, and include examples of how to use it?
- [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

- [x] Learning objectives: Does the module make the learning objectives plainly clear? (We don't require explicitly written learning objectives; only that they be evident from content and design.)
- [x] Content scope and length: Is the content substantial for learning a given topic? Is the length of the module appropriate?
- [x] Pedagogy: Does the module seem easy to follow? Does it observe guidance on cognitive load? (working memory limits of 7 +/- 2 chunks of information)
- [x] Content quality: Is the writing of good quality, concise, engaging? Are the code components well crafted? Does the module seem complete?
- [x] Instructional design: Is the instructional design deliberate and apparent? For example, exploit worked-example effects; effective multi-media use; low extraneous cognitive load.

- [x] Authors: Does the paper.md file include a list of authors with their affiliations?
- [x] A statement of need: Does the paper clearly state the need for this module and who the target audience is?
- [x] Description: Does the paper describe the learning materials and sequence?
- [x] References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
