+++ docs/reference-keys.txt
@@ -114,13 +114,17 @@ git-and-github
 the-three-ways
 set-up-your-github
 using-rstudio-with-git
+check-git-is-installed
+configure-your-git
+start-your-git
+connect-git-to-github
 using-the-git-gui---way-1
 create-a-new-version-control-in-rstudio---way-2
 if-have-have-an-existing-project---way-3
 commiting-to-git
-using-git-outside-rstudio
 push-to-github
 pull-from-github
+using-git-outside-rstudio
 troubleshooting
 were-you-challenged-for-your-password
 greyed-out-push-button
