+++ 04-prac4.Rmd
@@ -98,15 +98,31 @@ knitr::include_graphics('prac4_images/github.png')
 
 ### Using RStudio with Git
 
-Now, as I’ve mentioned before, RStudio is totally bad-ass. Not only does it make R fun to use, but the lovely people who created it also built in support for things like git!
+In summer 2021 GitHub changed it's authentication process to a token based system as opposed to a password based system. David Keys provided an [excellent overview with some videos that documented this change and how to now set things up, which i have adapted here](https://rfortherestofus.com/2021/02/how-to-use-git-github-with-r/)
 
-For a full and excellent tutorial on using Git with R Studio, watch [this webinar](https://rstudio.com/resources/webinars/managing-part-2-github-and-rstudio/)
+#### Check Git is installed 
 
-If you don’t want to watch the vid, I’ll do a quick summary below. So, to use git, first you need to enable it in RStudio:
+In the console window you will see a terminal tab, check Git is installed with `which git` then `git --version` you should get a message in response that says where your git installation is and the version you have 
 
-At the time of writing git integration *should* work within RStudio. If it doesn't try this again on your laptop.
+#### Configure your Git
 
-The next part of the practical is going run through the three ways of using Git and GitHub with RStudio as I laid out in [The three ways]
+You need to tell Git who you are and your GitHub username. The easiest way is to use the `usethis` package, you will need to install and library it.
+
+Then, in the console, type the function `edit_git_config()`
+
+A Git config will load and you need to change your name and email to match GitHub.
+
+#### Start your Git
+
+To start Git you **need to be in a RStudio project**. Instructions below show you how to do this in various scenarios. In the first instance we just copy it from the remote (GitHub) - so don't do this now....but a very simple way is to again load the `usethis` package in the console and the function `use_git()`, then type option 1. 
+
+#### Connect Git to GitHub
+
+From GitHub you need to generate a personal access token. You can use the function `create_github_token()` from the `usethis` package or also through GitHub > settings > Developer settings > personal access tokens > generate new token.
+
+Use a descriptive name and consider saving the token - it won't save on GitHub
+
+The last step is to store this token in Git with the `gitcreds` package > install and load it > then use the function `gitcreds_set()` > copy your token in.
 
 ### Using the Git GUI - way 1
 
@@ -144,7 +160,7 @@ knitr::include_graphics('prac4_images/existing.png')
 
 There is an easier way to set up Git and GitHub with your project, but this assumes you are starting fresh (with no code in an RProject)!
 
-1. Under [Set up your GitHub] we made a respository on GitHub. Copy that URL.
+1. Under [Set up your GitHub] we made a repository on GitHub. Copy that URL.
 
 1. Open RStudio > File New Project > Version Control > Git
 
@@ -156,8 +172,6 @@ knitr::include_graphics('prac4_images/easyway.png')
 
 ### If have have an existing project - way 3
 
-> Start with a fesh GitHub repository, we're assuming here that you have some code and then want to share it. DO NOT SELECT a README.md file...it should be an empty GitHub repo...
-
 1. Open RStudio and your existing project (or make a new one...i will make one here). In RStudio Tools > Global Options, under ‘Git/SVN’ check the box to allow version control and locate the folder on your computer where the git.exe file is located --- if you have installed git then this should be automatically there. If you make a new project make sure you create a file (`.R` or `.Rmd` through File > New File), add something to it, then save it (File > Save As) into your project folder. When it saves it should apprear in the bottom right Files window.
 
[…]
