+++ 03-prac3.Rmd
@@ -9,7 +9,7 @@ By the end of this practical you should be able to:
 
 ## Homework
 
-Outsides of our scheduled sessions you should be doing around 12 hours of extra study per week. Feel free to follow your own GIS interests, but good places to start include the following:
+Outside of our scheduled sessions you should be doing around 12 hours of extra study per week. Feel free to follow your own GIS interests, but good places to start include the following:
 
 ::: {.infobox .assignment data-latex="{note}"}
 **Exam**
