### 2021-11-17 #141 @darribas
I've just started with this and aiming to finish it today. A quick question out of my ignorance: I cannot find the PDF of the paper, have I missed it somehow? I can see the [`paper.md`](https://github.com/andrewmaclachlan/CASA0005repo/blob/master/paper.md) file but not the rendering version. From previous refereeing for JOSE, I seem to remember the issue used to include a link to it?
