### 2021-11-17 #141 @darribas
> Comments by @darribas

# Conflict of interest

- [x] As the reviewer I confirm that I have read the JOSE conflict of interest policy and that there are no conflicts of interest for me to review this work.

# Code of Conduct

- [x] I confirm that I read and will adhere to the JOSE code of conduct.

# General checks

> Overall, all good. A couple of (very) minor point from my side:
>
> - I didn't find a release tag that ensured the materials on the repository
>   were pinned to a version (e.g. `2.0`). It might be good for historical
>   reasons if someone would like to examine the version the paper is based on

- [x] Repository: Is the source for this learning module available at the repository url?
- [x] License: Does the repository contain a plain-text LICENSE file with the contents of a standard license? (OSI-approved for code, Creative Commons for content)
- [ ] Version: Does the release version given match the repository release (2.0)?

- [x] Authorship: Has the submitting author (@andrewmaclachlan) made visible contributions to the module? Does the full list of authors seem appropriate and complete?

> Again, this is very good in general. The statement of need in the paper
> makes a clear case. I wonder if parts of that blurb might even be ported to
> the "Welcome" or "Hello GIS" sections of the website.
>
> One point where I am surprised I didn't find much is on the installation of
> the stack. There's clear documentation to install RStudio but, in a course
> like this, it's only half the job. From my own teaching experience,
> installing some of the libraries for geospatial (e.g. `sf`, `rgdal`) can be
> rather tricky, and students usually run into issues if they already have
> some versions and need updates, etc. The book mentions an online alternative
> at UCL but, for non-UCL students I wonder if a bit more signposting could be
> added? Or at least pointing to external install guides that'd help them?
>
> On the topic on the stack, I tried several times to build the
> Binder instance and all resulted in failure to build. It is not critical
> but, if it doesn't work, is it work mentioning or removing?
>
> I did not find a clear document detailing community contributions. This may
> not be as important as for an open source library, but it might be nice to
> have something that tells people how you expect them to contribute if they
> want (e.g. submit issues, PRs, etc.). I have personally found very useful:
> https://github.com/UofTCoders/rcourse/blob/master/CONTRIBUTING.md

- [x] A statement of need: Do the authors clearly state the need for this module and who the target audience is?
- [ ] Installation instructions: Is there a clearly stated list of dependencies?
- [x] Usage: Does the documentation explain how someone would adopt the module, and include examples of how to use it?

- [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support

# Pedagogy / Instructional design (Work-in-progress: reviewers, please comment!)

> I absolutely _love_ the approach taken and the style. I think it is very
> conducive to build an inclusive environment that shows the students
> programming is not an obscure practice that requires lots of hard skills but
> is something everyone can try and enjoy. Plus, those music recommendations!
>
> If there is something I miss a bit on is more guidance for instructors who
> want to take the materials and make them their own, or teach off them. I
> don't think this needs to be changed for the paper to be accepted at all,
> I'm just suggesting it in case the authors consider it might be useful to
> have an additional section or notebook at the end geared more for
> instructors than students. I think it'd help the course gain (even) more
> traction.

- [x] Learning objectives: Does the module make the learning objectives plainly clear? (We don't require explicitly written learning objectives; only that they be evident from content and design.)
- [x] Content scope and length: Is the content substantial for learning a given topic? Is the length of the module appropriate?
- [x] Pedagogy: Does the module seem easy to follow? Does it observe guidance on cognitive load? (working memory limits of 7 +/- 2 chunks of information)
- [x] Content quality: Is the writing of good quality, concise, engaging? Are the code components well crafted? Does the module seem complete?

# Instructional design: Is the instructional design deliberate and apparent? For example, exploit worked-example effects; effective multi-media use; low extraneous cognitive load.

> See some of my comments above, I think this is one of the key strengths for
> the course.
>
> Another aspect I really enjoyed and I think contributes to the "inclusive
> vibe" of the course is the graphics.

# JOSE paper

- [x] Authors: Does the paper.md file include a list of authors with their affiliations?
- [x] A statement of need: Does the paper clearly state the need for this module and who the target audience is?
- [x] Description: Does the paper describe the learning materials and sequence?
- [x] References: Do all archival references that should have a DOI list one (e.g., papers, datasets, software)?
