+++ 07-prac7.Rmd
@@ -39,6 +39,10 @@ This week:
 * [Chapter 9 Spatial regression models](https://maczokni.github.io/crime_mapping_textbook/spatial-regression-models.html) from Crime Mapping in R by Juanjo Medina and Reka Solymosi (2019).
 
 Remember this is just a starting point, explore the [reading list](https://rl.talis.com/3/ucl/lists/139FBAF8-DACD-60FB-8BDC-E9C9E09BA885.html?lang=en-GB&login=1), practical and lecture for more ideas.
+
+**Doing**
+
+If you need a refresher/into to summary stats go through [Teacup Giraffes](https://tinystats.github.io/teacups-giraffes-and-statistics/02_bellCurve.html)
 :::
 
 ## Recommended listening `r emo::ji("headphones")`
@@ -382,6 +386,8 @@ ggplot(LonWardProfiles, aes(x=average_gcse_capped_point_scores_2014)) +
 
 ```
 
+Here, adding `..density..` means that the histogram is a density plot, this plots the chance that any value in the data is equal to that value.
+
 ```{r, message=FALSE, cache=TRUE}
 ggplot(LonWardProfiles, aes(x=unauthorised_absence_in_all_schools_percent_2013)) +
   geom_histogram(aes(y = ..density..),
@@ -464,6 +470,8 @@ qplot(x = log(median_house_price_2014),
       y = average_gcse_capped_point_scores_2014, 
       data=LonWardProfiles)
 ```
+
+Depending on if the independent or dependent (GCSE point score) variables have been transformed depends on how we interpret them - [see these rules for interpretation](https://data.library.virginia.edu/interpreting-log-transformations-in-a-linear-model/)
 
 #### Should I transform my variables?
 
@@ -629,6 +637,8 @@ plot(model2)
 
 In the series of plots above, the first plot (residuals vs fitted), we would hope to find a random cloud of points with a straight horizontal red line. Looking at the plot, the curved red line would suggest some hetroscedasticity, but the cloud looks quite random. Similarly we are looking for a random cloud of points with no apparent patterning or shape in the third plot of standardised residuals vs fitted values. Here, the cloud of points also looks fairly random, with perhaps some shaping indicated by the red line. 
 
+[Section 5.7.6 in Tidyverse Skills for Data Science explains each of these plots in more detail](https://jhudatascience.org/tidyversecourse/model.html#model-diagnostics)
+
 ### Assumption 5 - Independence of Errors
 
 This assumption simply states that the residual values (errors) in your model must not be correlated in any way. If they are, then they exhibit *autocorrelation* which suggests that something might be going on in the background that we have not sufficiently accounted for in our model. 
@@ -682,7 +692,9 @@ coordsW <- LonWardProfiles%>%
 
 plot(coordsW)
 
-#Now we need to generate a spatial weights matrix (remember from the lecture a couple of weeks ago). We'll start with a simple binary matrix of queen's case neighbours
+#Now we need to generate a spatial weights matrix 
+#(remember from the lecture a couple of weeks ago). 
+#We'll start with a simple binary matrix of queen's case neighbours
 
 LWard_nb <- LonWardProfiles %>%
   poly2nb(., queen=T)
@@ -705,13 +717,13 @@ plot(LWard_knn, st_geometry(coordsW), col="blue")
 #create a spatial weights matrix object from these weights
 
 Lward.queens_weight <- LWard_nb %>%
-  nb2listw(., style="C")
+  nb2listw(., style="W")
 
 Lward.knn_4_weight <- LWard_knn %>%
-  nb2listw(., style="C")
+  nb2listw(., style="W")
 ```
 
[…]
