+++ docs/reference-keys.txt
@@ -3,7 +3,7 @@ fig:unnamed-chunk-141
 fig:unnamed-chunk-147
 fig:unnamed-chunk-148
 fig:vectorplots
-fig:unnamed-chunk-429
+fig:unnamed-chunk-430
 geographic-information
 learning-outcomes
 homework
