+++ 06-prac6.Rmd
@@ -380,8 +380,13 @@ We can conduct a Ripley's K test on our data very simply with the `spatstat` pac
 K <- BluePlaquesSub.ppp %>%
   Kest(., correction="border") %>%
   plot()
+
+Kval <- as.data.frame(Kest(BluePlaquesSub.ppp, correction = "border"))
+
 ```
 The plot for K has a number of elements that are worth explaining. First, the *Kpois(r)* line in Red is the theoretical value of K for each distance window (r) under a Poisson assumption of Complete Spatial Randomness. The Black line is the estimated values of K accounting for the effects of the edge of the study area. 
+
+Here, the correction specifies how points towards the edge are dealt with, in this case, border means that points towards the edge are ignored for the calculation but are included for the central points. [Section 2.1, here](https://www.statistics.gov.hk/wsc/IPS031-P2-S.pdf) explains the different options.
 
 Where the value of K falls above the line, the data appear to be clustered at that distance. Where the value of K is below the line, the data are dispersed. From the graph, we can see that up until distances of around 1300 metres, Blue Plaques appear to be clustered in Harrow, however, at around 1500 m, the distribution appears random and then dispersed between about 1600 and 2100 metres. 
 
