+++ paper.md
@@ -22,6 +22,10 @@ bibliography: paper.bib
 
 This paper presents a masters level geographic information systems and science course (CASA0005) developed by the Bartlett Centre for Advanced Spatial Analysis (CASA), University College London (UCL). CASA0005 is a compulsory module for both MSc programs within the department, running annually in the autumn term with between 100 and 150 enrolled students. During the summer of 2019 the module was transformed from static practicals with some legacy content related to commercial software to an online interactive book primarily for the open source R data science programming language written using the Bookdown package [@Xie2016].
 
+## Resource story
+
+The geographic information systems and science course (CASA0005) is a core module on the CASA MSc and MRes programmes. Originally conceived primarily using ESRI's ArcMap with some Quantum GIS (QGIS), in recent years, consecutive module leaders and module staff, namely the authors of this paper, have transitioned and developed the content to R, with some QGIS. This reflects feedback from alumni, academia and industry that value programmatic and importantly reproducible GIS analysis. The course is representative of the expertise of the module staff with point pattern originating from Adam Dennett, raster analysis from Andrew MacLachlan and geographical regression a collaboration.   
+
 # Statement of need
 
 The buzzword ‘Big data’ and associated roles of data scientists have become common additions to academia, industry and government organisations seeking to optimise workflows and base decisions on factual evidence [@Donoho2017; @Li2016]. In response, the demand for data scientists has significantly grown, with two major recruitment websites LinkedIn and Glassdoor reporting a respective 56% and 44% year on year increase in advertised roles along with classification as the most promising career based on salary, growth, promotion opportunities and satisfaction within the USA in 2019 [@Glassdoor2020; @Pattabiraman2019]. In a similar theme data from the UK’s Universities and Colleges Admissions Service (UCAS) shows that the number of postgraduate level data science courses within the UK has grown from 104 across 51 institutions to 195 across 85 institutions between the academic years commencing in 2019 and 2020 [@TheUniversitiesandCollegesAdmissionsService2020]
