+++ README.md
@@ -2,19 +2,18 @@
 
 <img src="general_images/casa_logo.jpg" width="100" >
 
-  <!-- badges: start -->
-  [![Launch Rstudio Binder](http://mybinder.org/badge_logo.svg)](https://mybinder.org/v2/gh/andrewmaclachlan/CASA0005repo/master?urlpath=rstudio)
-  <!-- badges: end -->
-
-**Note** the binder for this book takes a long time to load and is mainly for demonstration purposes. For instructions on how to use binder please see the practical on Git, GitHub and RMarkdown.
+[![status](https://jose.theoj.org/papers/a05734775775d1ef968e582ce9a7cb17/status.svg)](https://jose.theoj.org/papers/a05734775775d1ef968e582ce9a7cb17)
 
 ## Description
+
+CASA0005 is an MSc level module created and delivered by Andrew MacLachlan and Adam Dennett from the Centre for Advanced Spatial Analysis (CASA), University College London (UCL).  
+
 
 This GitHub repository generates the CASA0005 Geographic Information Systems and Science student practical handbook which can be found at: https://andrewmaclachlan.github.io/CASA0005repo/
 
 ## File structure 
 
-Within each practical the assocaited data is stored in the prac(number)_data folder for example prac2_data. The image folders simply stored all the graphics used within the practical book to explain concepts/provide examples.
+Within each practical the associated data is stored in the prac(number)_data folder for example prac2_data. The image folders simply stored all the graphics used within the practical book to explain concepts/provide examples.
 
 ## Previous versions
 
