+++ README.md
@@ -6,14 +6,45 @@
 
 ## Description
 
-CASA0005 is an MSc level module created and delivered by Andrew MacLachlan and Adam Dennett from the Centre for Advanced Spatial Analysis (CASA), University College London (UCL).  
+CASA0005 is an MSc level module created and delivered by [Andrew MacLachlan](https://twitter.com/andymaclachlan) and [Adam Dennett](https://twitter.com/adam_dennett) from the [Centre for Advanced Spatial Analysis (CASA)](https://www.ucl.ac.uk/bartlett/casa/), University College London (UCL).  
 
+The module is a combination of lectures and practicals, with this repository holding the code to generate the student practical handbook, which can be found at: https://andrewmaclachlan.github.io/CASA0005repo
 
-This GitHub repository generates the CASA0005 Geographic Information Systems and Science student practical handbook which can be found at: https://andrewmaclachlan.github.io/CASA0005repo/
+The content covers the principles underlying the conception, representation/measurement and analysis of spatial phenomena. As such, it presents an overview of the core organising concepts and techniques of Geographic Information Systems, and the software and analysis systems that are integral to their effective deployment in spatial analysis. It is concerned with unearthing and understanding the importance of spatial data in a range of applied contexts.
 
-## File structure 
+## Target audience
 
-Within each practical the associated data is stored in the prac(number)_data folder for example prac2_data. The image folders simply stored all the graphics used within the practical book to explain concepts/provide examples.
+The practical material has primarily been developed for MSc (or MRes) level students who have little or no background in GIS, spatial analysis or the R data science language, but have studied in a related field such as: geography, architecture, computer science, statistics, mathematics or planning.
+
+However, there is no requirement to have any background in these disciplines to go through the material and anyone with an interest in spatial analysis in the R data science programming language will be able to successfully complete the practicals. 
+
+If you have any questions reach out to us on Twitter:
+
+* [Andy MacLachlan](https://twitter.com/andymaclachlan)
+* [Adam Dennett](https://twitter.com/adam_dennett)
+
+## Installation 
+
+### Students
+
+It is recommended that students follow the [software installation guide](https://andrewmaclachlan.github.io/CASA0005repo/software-installation.html) within the practical book and work through the material in their own RStudio instance. If they wish to replicate the practical book, see the next section.
+
+### Instructors 
+
+If instructors wish to replicate the practical book follow the guidance in [External usage](https://andrewmaclachlan.github.io/CASA0005repo/external-usage.html#external-usage) 
+
+## License
+
+Shield: [![CC BY-SA 4.0][cc-by-sa-shield]][cc-by-sa]
+
+This work is licensed under a
+[Creative Commons Attribution-ShareAlike 4.0 International License][cc-by-sa].
+
+[![CC BY-SA 4.0][cc-by-sa-image]][cc-by-sa]
+
+[cc-by-sa]: http://creativecommons.org/licenses/by-sa/4.0/
+[cc-by-sa-image]: https://licensebuttons.net/l/by-sa/4.0/88x31.png
+[cc-by-sa-shield]: https://img.shields.io/badge/License-CC%20BY--SA%204.0-lightgrey.svg
 
 ## Previous versions
 
@@ -36,7 +67,9 @@ Improvements made 2021 - 2022:
   * Instructions on citing papers in RMarkdown - this is now on the dissertation template. 
   * Interpolation from the first raster practical 
 1. Added more info on spatial weights and deciding between Spatial Error Models (SEMs) and Spatially Lagged Models (SLMs)
-1. Markscheme revision for an exam not assignment.
+1. Added more info on creating GWR bandwidths. 
+1. Mark scheme revision for an open book exam not written assignment.
+1. Github classroom for submission of the exam.
 
 Improvements made 2020 - 2021:
 
