+++ 01-prac1.Rmd
@@ -59,7 +59,7 @@ Some of these practicals are long, take regular breaks and have a listen to some
 
 Geographic data, geospatial data or geographic information is data that identifies the location of features on Earth. There are two main types of data which are used in GIS applications to represent the real world. **Vectors** that are composed of points, lines and polygons and **rasters** that are grids of cells with individual values...
 
-```{r echo=FALSE, out.width = "300pt", fig.align='center', cache=TRUE}
+```{r echo=FALSE, out.width = "300pt", fig.align='center', cache=TRUE, fig.cap="Types of spatial data. Source: [Spatial data models](https://planet.uwc.ac.za/nisl/gis/web_page/page_15.htm)"}
 knitr::include_graphics('prac1_images/rasvec.jpg') 
 ```
 
@@ -75,7 +75,7 @@ Discrete data have finite values. It can be numeric such as number of shoes or c
 
 Foot length would be continuous data but shoe size would be discrete data.
 
-```{r echo=FALSE, out.width = "450", fig.align='center', cache=TRUE}
+```{r echo=FALSE, out.width = "450", fig.align='center', cache=TRUE, fig.cap="Continuous and discrete data. Source: [Allison Horst data science and stats illustrations](https://github.com/allisonhorst/stats-illustrations)"}
 knitr::include_graphics('allisonhorst_images/continuous_discrete.png')
 ```
 
@@ -85,7 +85,7 @@ Ordinal data has an order or scale, for example if you have ever seen the chilli
 
 Binary data is that that can have only two possible outcomes, yes and no or shark and not shark.
 
-```{r echo=FALSE, out.width = "550pt", fig.align='center', cache=TRUE}
+```{r echo=FALSE, out.width = "550pt", fig.align='center', cache=TRUE, fig.cap="Nominal, ordinal and binary data. Source: [Allison Horst data science and stats illustrations](https://github.com/allisonhorst/stats-illustrations)"}
 knitr::include_graphics('allisonhorst_images/nominal_ordinal_binary.png')
 ```
 
@@ -154,7 +154,7 @@ knitr::include_graphics('prac1_images/geodatabase.png')
 
 #### GeoPackage
 
-```{r echo=FALSE, out.width = "100pt", fig.align='center', cache=TRUE}
+```{r echo=FALSE, out.width = "100pt", fig.align='center', cache=TRUE, fig.cap="GeoPacakge logo"}
 knitr::include_graphics('prac1_images/geopkg.png')
 ```
 
@@ -162,7 +162,7 @@ A GeoPackage is an open, standards-based, platform-independent, portable, self-d
 
 #### SpatiaLite
 
-```{r echo=FALSE, out.width = "100pt", fig.align='center', cache=TRUE}
+```{r echo=FALSE, out.width = "100pt", fig.align='center', cache=TRUE, fig.cap="SpatialLite logo"}
 knitr::include_graphics('prac1_images/spatialite.png')
 ```
 
@@ -170,7 +170,7 @@ SpatialLite is an open-source library that extends SQLite core. Support is fairl
 
 #### PostGIS
 
-```{r echo=FALSE, out.width = "100pt", fig.align='center', cache=TRUE}
+```{r echo=FALSE, out.width = "100pt", fig.align='center', cache=TRUE, fig.cap="PostGIS logo"}
 knitr::include_graphics('prac1_images/postGIS.jpg') 
 ```
 
@@ -186,13 +186,13 @@ As Grolemund and Wickham state in [R for Data Science](https://r4ds.had.co.nz/).
 
 > "Data science is a huge field, and there's no way you can master it by reading a single book."
 
[…]
