+++ docs/reference-keys.txt
@@ -1,8 +1,39 @@
+fig:unnamed-chunk-9
+fig:unnamed-chunk-10
+fig:unnamed-chunk-11
+fig:unnamed-chunk-14
+fig:unnamed-chunk-15
+fig:unnamed-chunk-16
+fig:unnamed-chunk-17
+fig:unnamed-chunk-18
+fig:unnamed-chunk-29
+fig:unnamed-chunk-32
+fig:unnamed-chunk-35
+fig:unnamed-chunk-73
+fig:unnamed-chunk-81
+fig:unnamed-chunk-92
+fig:unnamed-chunk-93
+fig:unnamed-chunk-94
+fig:unnamed-chunk-105
+fig:unnamed-chunk-115
+fig:unnamed-chunk-117
+fig:unnamed-chunk-120
+fig:unnamed-chunk-123
+fig:unnamed-chunk-126
+fig:unnamed-chunk-133
+fig:unnamed-chunk-138
 fig:unnamed-chunk-139
 fig:unnamed-chunk-141
 fig:unnamed-chunk-147
 fig:unnamed-chunk-148
 fig:vectorplots
+fig:unnamed-chunk-176
+fig:unnamed-chunk-343
+fig:unnamed-chunk-344
+fig:unnamed-chunk-345
+fig:unnamed-chunk-346
+fig:unnamed-chunk-347
+fig:unnamed-chunk-348
 fig:unnamed-chunk-430
 geographic-information
 learning-outcomes

[…]
