### 2022-01-13 #141 @andrewmaclachlan
Hi there, please see the response to @TomasBeuzen's review below. @darribas gave a few similar comments, but i'll address those in full over the next few days as well.

- [ ] The only thing missing is a reference to what the author's expertise is in the area, which is one of the review checklist items (Does the paper tell the "story" of how the authors came to develop it, or what their expertise is?). Not sure how important that is to have though and it didn't bother me as the quality of the material speaks for itself.
- The paper now has a resource story section
- [ ] The README is the first place I'll often visit when looking at other's material. I think it needs to include at least a brief paragraph (1-2 sentences) about what the material covers and who the target audience is.
- The README has been updated with a description, target audience, installation, license and contribution sections
- [ ] I would like to see clear instructions for how to install the material locally and develop it. Imagine someone wanting to use it with minimal experience with GitHub/Bookdown/etc, e.g., 1. Clone this repository git clone git@github.com:andrewmaclachlan/CASA0005repo.git. 2. Install R and RStudio. 3. Install Bookdown.... Or more helpful would be a Docker image - I wouldn't require it to pass this review but having personally worked with bookdown to build course materials like this, it's a life-saver for your future self and others trying to use your material, so worth considering.
- I did create a binder, more as an experiment, but as the material is rather large it would take a long time to load and in some instances wouldn't - i think Dani experienced this. As a result i've removed it and instead added a section on the bookdown website about how to adpot the course with specific instructions that include a renv snapshot which will install the correct version of the packages used. https://andrewmaclachlan.github.io/CASA0005repo/external-usage.html#how-to-adopt-this-course
- [ ] There are no contributing guidelines, as far as I could see. I note "Feedback" forms in the material itself. But I think some guidelines in the README of main repo are needed, i.e. are you accepting PR's or issues? If so, how would you like them structured, or what kind of things can someone contribute?
- I have added a how to contribute section in the bookdown website, and the README links to it - https://andrewmaclachlan.github.io/CASA0005repo/external-usage.html#how-to-contribute.
- [ ] There's an MIT license for the repo. It gets a little grey when sharing a repo with both original code and content. I usually associate an MIT license with software, but not necessarily written content, figures, etc. A CC license might be a more suitable option here? Or worth considering as an additional license that covers the content. Apologies if this is already covered somewhere and I missed it.
-  I've added a badge and licence section on the REAME, included a license file (see the about section) and updated the bookdown website with the same license.
 - [ ] Figures throughout the material are missing captions. In particular, I'm missing references to original sources. I note that you acknowledge these various sources in the welcome page, but as readers may not read that page, I think it's appropriate to add references to original figures when they are used in the text.
- I've added figure captions under each figure where they have been taken from an original source.

- [ ] License: Does the repository contain a plain-text LICENSE file with the contents of a standard license? (OSI-approved for code, Creative Commons for content)
- See comments above, i have added this to the repo. https://github.com/andrewmaclachlan/CASA0005repo/blob/master/LICENSE

-[ ] A statement of need: Do the authors clearly state the need for this module and who the target audience is?
 Installation instructions: Is there a clearly stated list of dependencies?
- There is a statement of need in the paper and i've added a target audience section to the README. The package section in the bookdown website holds all the package info, that can be replicated with renv
- [ ] Usage: Does the documentation explain how someone would adopt the module, and include examples of how to use it?
- Addressed above, yes more specific adoption instructions
- [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support
- Yes, clear contribution guidance is provided on the bookdown website - see link above.

- Yes, i've added this to the paper.
