### 2022-01-13 c8abd81
few minor edits for the paper
