+++ dani_response.md
@@ -1,0 +1,14 @@
+### General checks
+- [ ]  Version: Does the release version given match the repository release (2.0)?
+- I did create the version myself which probably wasn't the best word to use. Version implies software to me, but i wanted to show and document how the content has been developed over the years. To rectify this i've added an updates section to the paper and README which reflects my intention for doing it in the manner I have. I've also changed the version section on the bookdown website to Updates.
+
+### Documentation 
+ - The binder instance was more of test of concept for me, however, i see your point so have removed it. Instead there is an renv snapshot which holds all the package information in and instructions to load it in: https://andrewmaclachlan.github.io/CASA0005repo/external-usage.html#how-to-adopt-this-course 
+ - [ ] Installation instructions: Is there a clearly stated list of dependencies?
+- There is a list of packages used on the bookdown website and also now the renv snapshot. 
+- [ ] Community guidelines: Are there clear guidelines for third parties wishing to 1) Contribute to the module 2) Report issues or problems with the module 3) Seek support
+- Yes, this was also picked up by Tomas. I've added a section on how to contribute following your suggestion. The README links to this section too: https://andrewmaclachlan.github.io/CASA0005repo/external-usage.html#how-to-contribute
+
+### Pedagogy / Instructional design
+- Tomas also commented about guidance for taking the materials are using them. I've added a section called how to adpot this course in the bookdown website. The README also links to this for students and for instructors. 
+  
