+++ paper.bib
@@ -101,7 +101,7 @@ year = {2017}
 author = {Glassdoor},
 booktitle = {Glassdoor},
 title = {{50 Best Jobs in America for 2019}},
-url = {https://www.glassdoor.com/List/Best-Jobs-in-America-2019-LST{\_}KQ0,25.htm},
+url = {https://www.glassdoor.co.uk/List/Best-Jobs-in-America-LST_KQ0,20.htm},
 urldate = {2020-04-07},
 year = {2020}
 }
