### 2022-02-01 #141 @TomasBeuzen
Thanks for addressing all my comments @andrewmaclachlan and congrats again on putting together a great course!

@ttimbers - I'm happy that my review has been addressed 👍 🚀
