### 2022-03-22 #141 @ttimbers
@openjournals/jose-eics - I ran `@whedon recommend-accept` but it seems stuck. It did not generate the PDF nor check the DOI's. Could you please provide some guidance as to how I should proceed here? Many thanks!
