### 2022-03-23 #141 @labarba
I see that @ttimbers added a checklist for the second reviewer in a separate [comment](https://github.com/openjournals/jose-reviews/issues/141#issuecomment-937324211). Is @darribas able to check off items there?

Normally the checklist is added automatically by whedon at the top of the thread when starting the review. I checked the Pre-Review issue to figure out why that didn't happen here. Bear in mind @ttimbers is a new editor. I think that [running](https://github.com/openjournals/jose-reviews/issues/85#issuecomment-937321982) `@whedon assign TomasBeuzen as reviewer` actually swapped the original assignment and we needed `@whedon add TomasBeuzen as reviewer` to keep `@darribas`. But I'm not sure because I see both reviewers listed at the top of the Pre-Review.
