### 2022-03-23 #141 @labarba
Ah! @ttimbers, I see you manually edited the reviews in the [top comment](https://github.com/openjournals/jose-reviews/issues/85#issue-660300287) of the Pre-Review. I think the `@whedon add` command is what we needed here.
