### 2022-03-25 #141 @labarba
@darribas Can you check off items in this [checklist](https://github.com/openjournals/jose-reviews/issues/141#issuecomment-937324211)?
