### 2022-04-04 #141 @ttimbers
@darribas - a gentle ping to follow-up on checking off the items in this checklist: https://github.com/openjournals/jose-reviews/issues/141#issuecomment-937324211
