### 2022-04-08 #141 @darribas
Hello, sorry for the slow movement on this front. I've checked all my original comments and @andrewmaclachlan answers and I'm happy with what it looks like now. I've ticked the missing boxes on the [original comment](https://github.com/openjournals/jose-reviews/issues/141#issuecomment-972204803) but not sure if I need to do anything else, let me know and I'll go ahead. Congratulations @andrewmaclachlan on the course, it's fantastic!!!
