### 2022-04-08 #141 @labarba
@ttimbers @darribas — I cleaned up this review issue a little bit. There seemed to be some confusion as to where the reviewer should tick off items in their review checklist. When things start smoothly from the Pre-review issue, each reviewer gets a checklist in the top-most comment of the Review thread. If one reviewer is added later, we add the checklist manually by editing that top comment. Here, a checklist had been added as a separate comment down the thread, which required lots of scrolling and things getting lost. Thank you for your patience!
