### 2022-04-09 #141 @labarba
I notice that the tag in the release is `jose`: https://github.com/andrewmaclachlan/CASA0005repo/tags
…whereas the version here was set as `2.0`. These two should match. Which one do you prefer, @andrewmaclachlan?
