### 2022-04-11 #141 @andrewmaclachlan
I changed the tagged release to 3.0 as in the text it now says: This is the third iteration of the practical book - https://andrewmaclachlan.github.io/CASA0005repo/external-usage.html
